### 2023-12-04 acb33e7
refs 82 fix DOI and clean up caption
