+++ paper/paper.bib
@@ -12,7 +12,7 @@
 	 year = {1992},
 	 note = {Quasielastic Neutron Scattering},
 	 issn = {0921-4526},
-	 doi = {https://doi.org/10.1016/0921-4526(92)90036-R},
+	 doi = {10.1016/0921-4526(92)90036-R},
 	 url = {https://www.sciencedirect.com/science/article/pii/092145269290036R},
 	 author = {D.S. Sivia and C.J. Carlile and W.S. Howells and S. König},
 	 abstract = {We consider the analysis of quasielastic neutron scattering data from a Bayesian point-of-view. This enables us to use probability theory to assess how many quasielastic components there is most evidence for in the data, as well as providing an optimal estimate of their parameters. We review the theory briefly, describe an efficient algorithm for its implementation and illustrate its use with both simulated and real data.}
