### 2023-12-04 c72b5b1
refs 82 fix DOI and clean up caption
