+++ paper/paper.md
@@ -29,7 +29,7 @@ The `quickBayes` package comes with an API that can be easily be extended by use
 
 ![ Two plots of the same raw data from a Quasi Elastic Neutron Scattering (QENS) experiment, showing the fits for one and two peaks (it also includes a linear background and an elastic peak).
 The inserts show zoomed in images of the peak centre.
-From the loglikelihood calculation the most likely number of peaks is two. \label{fig:peaks} ] (figures/peaks.png) {width=80%} {height="22pt"}
+From the loglikelihood calculation the most likely number of peaks is two. \label{fig:peaks} ](figures/peaks.png){width=80%}{height="22pt"}
 
 
 # Statement of need
