### 2024-01-07 ISISNeutronMuon/quickBayes#86 @osorensen
Editorial pre-review comments on JOSS submission
I am the editor of your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/6106). Here are some suggested corrections based on my first reading of your submission.

### Comment 1

- [ ] On line 7, change from "hypothesis" to "hypotheses", since it's plural:

### Comment 2

- [ ] On line 10, make sure the reference to the figure compiles properly:

### Comment 3

- [ ] On line 14, make sure to capitalize Lorentzian:

### Comment 4

- [ ] On line 18, remove the formatting text:

### Comment 5

There must be a typo in the title of this book:

In addition, I was not able to find it either on Amazon or by searching Google. Could you please add the ISBN of the book to the reference?
