### 2024-01-08 f13e63d
refs #82 paper changes from review
