+++ paper/paper.bib
@@ -37,6 +37,6 @@
       author = {Devinder Silvia, John Skilling},
       edition = {Second},
       publisher = {Oxford University Press},
-      title = {Data Analysis A Bayesoam Tutorial},
+      title = {Data Analysis A Bayesiam Tutorial},
       year = {2006}
 }
