### 2024-01-08 1ca9cf9
refs #82 paper changes from review
