+++ paper/paper.bib
@@ -38,5 +38,6 @@
       edition = {Second},
       publisher = {Oxford University Press},
       title = {Data Analysis A Bayesiam Tutorial},
-      year = {2006}
+      year = {2006},
+      isbn = {9780198568322}
 }
