+++ paper/paper.bib
@@ -37,7 +37,7 @@
       author = {Devinder Silvia, John Skilling},
       edition = {Second},
       publisher = {Oxford University Press},
-      title = {Data Analysis A Bayesiam Tutorial},
+      title = {Data Analysis A Bayesian Tutorial},
       year = {2006},
       isbn = {9780198568322}
 }
