### 2024-01-15 #6106 @JonasMoss
Sure, I'd be happy to.

On Sat, Jan 13, 2024 at 8:57 PM Øystein Sørensen ***@***.***>
wrote:

> 👋 @JonasMoss <https://github.com/JonasMoss>, @jdrugo
> <https://github.com/jdrugo>, @hiendn <https://github.com/hiendn>, would
> any of you be willing to review this submission for JOSS? We carry out our
> checklist-driven reviews here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
