### 2024-01-15 #6230 @editorialbot
[REVIEW]: quickBayes: An analytical approach to Bayesian loglikelihoods
**Submitting author:** @AnthonyLim23 (<a href="http://orcid.org/0009-0008-2594-0652">Anthony Lim</a>)
**Repository:** https://github.com/ISISNeutronMuon/quickBayes
**Branch with paper.md** (empty if default branch): 82_paper
**Version:** 1.0.0b18
