### 2024-01-16 #6230 @mattpitkin
Hi @AnthonyLim23, I've just got a few minor things to suggest initially:

* make the links to the documentation in the GitHub [README file](https://github.com/ISISNeutronMuon/quickBayes/blob/main/README.md) into actual working links (just switch from `rst` style links to Markdown style);
* add the installation instructions to the documentation in https://quickbayes.readthedocs.io/;
* fix the [`paper.bib`](https://github.com/ISISNeutronMuon/quickBayes/blob/82_paper/paper/paper.bib) file, so that the authors display properly in the paper, i.e., change it to:

```bib
@software{quasielasticbayes,
	  author = {{Sivia}, D. and {Howells}, S.},
	  title = {quasielasticbayes version 0.2.0}
}

@article{bayesPaper,
	 title = {Bayesian analysis of quasielastic neutron scattering data},
	 journal = {Physica B: Condensed Matter},
	 volume = {182},
	 number = {4},
	 pages = {341-348},
	 year = {1992},
	 note = {Quasielastic Neutron Scattering},
	 issn = {0921-4526},
	 doi = {10.1016/0921-4526(92)90036-R},
	 url = {https://www.sciencedirect.com/science/article/pii/092145269290036R},
	 author = {{Sivia}, D.~S. and {Carlile} C.~J. and {Howells}, W.~S. and {König}, S.},
	 abstract = {We consider the analysis of quasielastic neutron scattering data from a Bayesian point-of-view. This enables us to use probability theory to assess how many quasielastic components there is most evidence for in the data, as well as providing an optimal estimate of their parameters. We review the theory briefly, describe an efficient algorithm for its implementation and illustrate its use with both simulated and real data.}
}
}

@article{bayesReview,
         title = {Bayesian inference in physics},
         author = {{von Toussaint}, Udo},
         journal = {Rev. Mod. Phys.},
         volume = {83},
         issue = {3},
         pages = {943--999},
         year = {2011},
         month = {Sep},
         publisher = {American Physical Society},
         doi = {10.1103/RevModPhys.83.943},
         url = {https://link.aps.org/doi/10.1103/RevModPhys.83.943}
}

@book{bayesBook,
      author = {{Sivia}, D. and {Skilling}, J.},
      edition = {Second},
      publisher = {Oxford University Press},
      title = {Data Analysis A Bayesian Tutorial},
      year = {2006},
      isbn = {9780198568322}
}
```
