### 2024-01-16 #6230 @mattpitkin
In the [developer](https://quickbayes.readthedocs.io/en/latest/dev/setup.html) section of the docs, could you please add a sentences or two stating how users/developers can:

> 2) Report issues or problems with the software 3) Seek support

i.e., just say that users can submit issues in the GitHub repo.
