### 2024-01-16 ISISNeutronMuon/quickBayes#87 @AnthonyLim23
Review feedback
The recommended changes from https://github.com/openjournals/joss-reviews/issues/6230#issuecomment-1893590813

- [x] add section in docs on reporting issues and getting support (github)
- [x] Fix links in Readme to work correctly (change to markdown style)
- [x] Add installation instructions to docs
