### 2024-01-16 854e51b
refs #82 update bib file
