+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @software{quasielasticbayes,
-	  author = {Devinder Silvia, Spencer Howells},
+	  author = {{Sivia}, D. and {Howells}, S.},
 	  title = {quasielasticbayes version 0.2.0}
 }
 
@@ -14,14 +14,14 @@
 	 issn = {0921-4526},
 	 doi = {10.1016/0921-4526(92)90036-R},
 	 url = {https://www.sciencedirect.com/science/article/pii/092145269290036R},
-	 author = {D.S. Sivia and C.J. Carlile and W.S. Howells and S. König},
+	 author = {{Sivia}, D.~S. and {Carlile} C.~J. and {Howells}, W.~S. and {König}, S.},
 	 abstract = {We consider the analysis of quasielastic neutron scattering data from a Bayesian point-of-view. This enables us to use probability theory to assess how many quasielastic components there is most evidence for in the data, as well as providing an optimal estimate of their parameters. We review the theory briefly, describe an efficient algorithm for its implementation and illustrate its use with both simulated and real data.}
 }
 }
 
 @article{bayesReview,
          title = {Bayesian inference in physics},
-         author = {von Toussaint, Udo},
+         author = {{von Toussaint}, Udo},
          journal = {Rev. Mod. Phys.},
          volume = {83},
          issue = {3},
@@ -34,7 +34,7 @@
 }
 
 @book{bayesBook,
-      author = {Devinder Silvia, John Skilling},
+      author = {{Sivia}, D. and {Skilling}, J.},
       edition = {Second},
       publisher = {Oxford University Press},
       title = {Data Analysis A Bayesian Tutorial},
