### 2024-01-16 b49e14f
refs #82 update bib file
