+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @software{quasielasticbayes,
-	  author = {{Sivia}, D. and {Howells}, S.},
+	  author = {{Sivia}, D.~S. and {Howells}, W.~S.},
 	  title = {quasielasticbayes version 0.2.0}
 }
 
@@ -34,7 +34,7 @@
 }
 
 @book{bayesBook,
-      author = {{Sivia}, D. and {Skilling}, J.},
+      author = {{Sivia}, D.~S. and {Skilling}, J.},
       edition = {Second},
       publisher = {Oxford University Press},
       title = {Data Analysis A Bayesian Tutorial},
