### 2024-01-16 81d20f4
refs #82 update bib file
