### 2024-01-16 42f2bc5
refs #87 doc tidy up
