### 2024-01-16 eeaf0e8
refs #87 doc tidy up
