+++ docs/source/install.rst
@@ -1,25 +1,19 @@
 Installation
 ============
 
-This package provides a convenient API
-for performing Bayesian analysis
-on muon and neutron-scattering data.
+The library is available on :ref:`PyPi<https://pypi.org/project/quickBayes/#description>` and can be installed with pip:
 
-## Installing
+.. code-block:: python
 
-The library is available on [PyPi](https://pypi.org/project/quickBayes/#description) and can be installed with pip:
-
-```sh
-> python -m pip install quickBayes
-```
+    python -m pip install quickBayes
 
 Reporting Issues
 ================
 
-To report an issue please create a [github issue](https://github.com/ISISNeutronMuon/quickBayes/issues) with the details of the bug.
+To report an issue please create a :ref:`github issue<https://github.com/ISISNeutronMuon/quickBayes/issues>` with the details of the bug.
 
 
 Support
 =======
 
-Support will be provided via [github issues](https://github.com/ISISNeutronMuon/quickBayes/issues), please outline the problem in a new issue.
+Support will be provided via :ref:`github issues<https://github.com/ISISNeutronMuon/quickBayes/issues>`, please outline the problem in a new issue.
