### 2024-01-16 488e262
refs #87 doc tidy up
