+++ docs/source/install.rst
@@ -1,7 +1,7 @@
 Installation
 ============
 
-The library is available on :ref:`PyPi<https://pypi.org/project/quickBayes/#description>` and can be installed with pip:
+The library is available on `PyPi<https://pypi.org/project/quickBayes/#description>` and can be installed with pip:
 
 .. code-block:: python
 
@@ -10,10 +10,10 @@ The library is available on :ref:`PyPi<https://pypi.org/project/quickBayes/#desc
 Reporting Issues
 ================
 
-To report an issue please create a :ref:`github issue<https://github.com/ISISNeutronMuon/quickBayes/issues>` with the details of the bug.
+To report an issue please create a `github issue<https://github.com/ISISNeutronMuon/quickBayes/issues>` with the details of the bug.
 
 
 Support
 =======
 
-Support will be provided via :ref:`github issues<https://github.com/ISISNeutronMuon/quickBayes/issues>`, please outline the problem in a new issue.
+Support will be provided via `github issues<https://github.com/ISISNeutronMuon/quickBayes/issues>`, please outline the problem in a new issue.
