### 2024-01-16 #6230 @AnthonyLim23
I have updated the PR for the paper with your bib file. I then edited the authors list for the software so the shared names are the same as the paper.

The other comments have been addressed in https://github.com/ISISNeutronMuon/quickBayes/pull/88
