### 2024-01-16 #6230 @mattpitkin
@AnthonyLim23 It looks like I made a typo in my version of the `.bib` file and `{Carlile} C.~J.` is missing a comma, so should be `{Carlile}, C.~J.`. Also, is there any possible date and URL that could be use for the `Quasielasticbayes v0.2.0` reference?

While, I'm at it - I spotted a minor typo. In the acknowledgements `quickBayes` is spelled wrong.
