### 2024-01-18 #6230 @AnthonyLim23
For the Quasielasticbayes 0.2.0 I dont have a date or url. It was a programme written late 80's early 90's for internal use by scientists. So was never properly documented. I can put the github that it has since been moved to (and updated to work with  Python), what do you think?
