+++ paper/paper.bib
@@ -1,6 +1,7 @@
 @software{quasielasticbayes,
 	  author = {{Sivia}, D.~S. and {Howells}, W.~S.},
-	  title = {quasielasticbayes version 0.2.0}
+	  title = {quasielasticbayes version 0.2.0},
+	  url = {https://github.com/mantidproject/quasielasticbayes}
 }
 
 @article{bayesPaper,
@@ -14,7 +15,7 @@
 	 issn = {0921-4526},
 	 doi = {10.1016/0921-4526(92)90036-R},
 	 url = {https://www.sciencedirect.com/science/article/pii/092145269290036R},
-	 author = {{Sivia}, D.~S. and {Carlile} C.~J. and {Howells}, W.~S. and {König}, S.},
+	 author = {{Sivia}, D.~S. and {Carlile}, C.~J. and {Howells}, W.~S. and {König}, S.},
 	 abstract = {We consider the analysis of quasielastic neutron scattering data from a Bayesian point-of-view. This enables us to use probability theory to assess how many quasielastic components there is most evidence for in the data, as well as providing an optimal estimate of their parameters. We review the theory briefly, describe an efficient algorithm for its implementation and illustrate its use with both simulated and real data.}
 }
 
