### 2024-02-19 #6230 @mattpitkin
Hi @AnthonyLim23, sorry for the slow response.

>  @mattpitkin for the API examples would you like to see examples for the individual pieces (e.g. fitting functions, fitting engines) or just a few example workflows/model selection? Most of the API docs are autogenerated. I did try to write the docs in the style of a tutorial, but I assume that those examples are insufficient. What changes would you recommend that I make to those examples to improve the API documentation?

I think it would be good to see, say, two _fully_ worked examples demonstrating a realistic pipeline of usage of the code, including either real or simulated datasets. One could be, say, determining the number of Gaussian-like "lines" in some data and one could also incorporated some modelling of a background (or whatever you think appropriate). If there's a real analysis that it has been used on where the data is open and you a free to post in the docs that would be great, e.g., the QENS example that you have in the paper. You could do these in examples in Jupyter notebooks and either just link to them in the docs or get [Sphinx to render them](https://nbsphinx.readthedocs.io/en/0.9.3/), so that you can easily include figures (although this is not a requirement).

Some other bits of the API do need work though. The [fit functions](https://quickbayes.readthedocs.io/en/latest/generated/quickBayes.fit_functions.htm) all at least need a short doc string saying what they are and what input parameters they require, otherwise they are not really useful.

In terms of other general comments (and again in relation to @JonasMoss's comments), I think the first paragraph of the paper nicely described the use case of the code, so that should be echoed in the docs. Essentially this code is for model comparison and for each model you are (I think/assume) approximating the [marginalised likelihood](https://en.wikipedia.org/wiki/Marginal_likelihood) for a given model (also known as the evidence), where you marginalising over all the unknown model parameters (can you hold an of the parameters at fixed values or do you have to marginalise over them all?) as opposed to an analytical maximum likelihood approach. Maybe this is done using [Laplace's](https://en.wikipedia.org/wiki/Laplace%27s_approximation) [method](https://www.inference.org.uk/mackay/itprnn/ps/341.342.pdf)? You can state this in the docs and paper, but you should also briefly mention other approximation methods such as the [AIC](https://en.wikipedia.org/wiki/Akaike_information_criterion), [BIC](https://en.wikipedia.org/wiki/Bayesian_information_criterion) and [nested sampling](https://en.wikipedia.org/wiki/Nested_sampling_algorithm) (which is the Monte Carlo method described at the end of the Sivia book). So, when you talk about loglikelihood in the docs I think you really mean marginal likelihood/evidence - please check if this is what you mean and make changes. You should also be explicit that priors on all model parameters are assumed to be flat and infinite in bounds (unless they are not?). This effects how you interpret the model comparison, so it does need to be stated. Is there any way to specifically set the prior ranges, or use different priors (say Gaussian priors, which can also be analytically marginalised over in certain cases)?

In the docs, it currently states that:

> "Fit functions are the mathematical representations of the data"

but I don't really think this is quite what they are. The fit functions are a (potentially approximate) mathematical model for a  process/phenomenon that you have collected data to measure. The data will be a combination of that modelled process and some noise (where the noise could just be measurement noise, or a combination of other unmodelled processes). So, you could say something like: "The fit functions are approximate mathematical models of the process that we intend to measure.".
