### 2024-02-19 #6230 @AnthonyLim23
@mattpitkin I have data in the tests for QENS and simulated muon data (number of exp decays). I can do two examples using them.

I will look into the rest of the API docs and make them clearer.

I believe that it is the marginalised likelihood, but I will double check. At presents all of the parameters have to be free, I do want to make it easier to add fixes (the current method is https://github.com/ISISNeutronMuon/quickBayes/blob/main/src/quickBayes/fit_functions/stretch_exp_fixed.py which is a bit clunky.

In terms of priors, from a user perspective it is purely the parameter range and start values for fit parameters. I believe the flat priors are baked into the assumptions used for deriving the analytical equation. I can touch on it, but the user will never be able to do anything about it.
