### 2024-02-27 ISISNeutronMuon/quickBayes#89 @prashjet
JOSS review: installation error
In a fresh conda environment, I get the following error when attempting the `pip` install as instructed on the `README`:

```
(quickBayes) prash@air > python -m pip install quickBayes
Collecting quickBayes
  Using cached quickBayes-1.0.0b18.tar.gz (1.6 MB)
  Preparing metadata (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py egg_info did not run successfully.
  │ exit code: 1
  ╰─> [6 lines of output]
      Traceback (most recent call last):
        File "<string>", line 2, in <module>
        File "<pip-setuptools-caller>", line 34, in <module>
        File "/private/var/folders/n2/qrrgr98x5flg7lx88y_r9z0r0000gn/T/pip-install-zwi18xxa/quickbayes_4aadff4f9d2b4124af579c739b4f2dcb/setup.py", line 2, in <module>
          from tools.setup_helper import get_extensions
      ModuleNotFoundError: No module named 'tools.setup_helper'
      [end of output]

error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.
```
