### 2024-02-27 #6230 @prashjet
Hi there. Thanks for your patience waiting for my review, I've just got round to it after a busy period.

After reading the documentation and paper, I am still quite confused about which calculations this software is doing, and what information those calculations are telling me. I think the [comment](https://github.com/openjournals/joss-reviews/issues/6230#issuecomment-1930099874) from @JonasMoss provides an excellent checklist of what should be included in the documentation so that someone can clearly understand what your software does and when to use it. I won't repeat the entire checklist here, but I'll highlight some key points in a response to your last message. You say...

> This package is a bit strange as it uses some Bayesian statistics, but not directly. So I want to check how much detail on Bayesian statistics is appropriate.

In principle, I think that  _black box_ software for Bayesian statistics may sometimes be useful (i.e. software that provides a solution, without necessarily going into details) but only if you can clearly explain what situations the black-box is suitable for. In the case of `quickBayes`, to answer this question you'll need to understand Equation 4.20 of the Sivia & Skilling. This will require:
- explaining all of the terms in Equation 4.20,
- explaining the assumptions that have gone into deriving Equation 4.20,
- doing both of the above using language/notation that is consistent throughout your software package.

That last point is a little opaque, so I'll try to clarify it with an example. Currently, your `loglikelihood` function takes an argument `N_peaks`. In the code snippet on [loglikelihood](https://quickbayes.readthedocs.io/en/latest/loglikelihood.html) page, you set `N_peaks=1` for both examples of `fit_functions`, even though one of them is a model with a single Gaussian, and one with two Gaussians. This leaves me asking what is the meaning of `N_peak` and how (if it all?) does it relate to the number of peaks I have defined in my `fit_function`?

I think unravelling Equation 4.20 of the Sivia & Skilling is likely to be difficult task, however, given that this calculation is currently at the heart of your software package, I think it is unavoidable in order for anyone know if `quickBayes` is the right tool for them. I think this will involve an entire re-write of the documentation.

p.s. I also had a problem with [installation](https://github.com/ISISNeutronMuon/quickBayes/issues/89#issue-2156887601).
