### 2024-02-29 e7843f9
Doc tidy up (#88)
* refs #87 doc tidy up

* refs #87 doc tidy up

* refs #87 doc tidy up
