+++ docs/source/install.rst
@@ -1,0 +1,19 @@
+Installation
+============
+
+The library is available on `PyPi<https://pypi.org/project/quickBayes/#description>` and can be installed with pip:
+
+.. code-block:: python
+
+    python -m pip install quickBayes
+
+Reporting Issues
+================
+
+To report an issue please create a `github issue<https://github.com/ISISNeutronMuon/quickBayes/issues>` with the details of the bug.
+
+
+Support
+=======
+
+Support will be provided via `github issues<https://github.com/ISISNeutronMuon/quickBayes/issues>`, please outline the problem in a new issue.
