### 2024-02-29 d885a7c
#refs 89 first go at fixing pip install
