+++ .github/workflows/release.yml
@@ -4,6 +4,12 @@ on:
   release:
     types:
         - published
+  push:
+    branches:
+    - 'main'
+  pull_request:
+    branches:
+      - 'main'
 
 jobs:
   build_wheels:
