### 2024-02-29 4057435
#refs 89 first go at fixing pip install
