### 2024-02-29 fb422c7
#refs 89 first go at fixing pip install
