### 2024-02-29 f481940
#refs 89 first go at fixing pip install
