### 2024-02-29 39d78e1
#refs 89 first go at fixing pip install
