### 2024-02-29 9f2f4a1
#refs 89 first go at fixing pip install
