### 2024-02-29 e6bff1f
#refs 89 first go at fixing pip install
