### 2024-02-29 52fcd4e
#refs 89 first go at fixing pip install
