### 2024-02-29 ba7863e
#refs 89 first go at fixing pip install
