### 2024-02-29 f10aabc
#refs 89 first go at fixing pip install
