+++ .github/workflows/release.yml
@@ -2,6 +2,8 @@ name: Build and upload to PyPI
 
 on:
   release:
+    types:
+        - published
 
 jobs:
   build_wheels:
