### 2024-02-29 ISISNeutronMuon/quickBayes#89 @AnthonyLim23
If you try a pip install now it should work
