### 2024-02-29 ISISNeutronMuon/quickBayes#89 @prashjet
Same error as before unfortunately:

```
(quickBayes) prash@air > python -m pip install quickBayes
Collecting quickBayes
  Downloading quickBayes-1.0.0b21.tar.gz (2.2 MB)
  Preparing metadata (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py egg_info did not run successfully.
  │ exit code: 1
  ╰─> [6 lines of output]
      Traceback (most recent call last):
        File "<string>", line 2, in <module>
        File "<pip-setuptools-caller>", line 34, in <module>
        File "/private/var/folders/n2/qrrgr98x5flg7lx88y_r9z0r0000gn/T/pip-install-b6yahjlx/quickbayes_2c34791c0c44403b97a236e18f59369e/setup.py", line 2, in <module>
          from tools.setup_helper import get_extensions
      ModuleNotFoundError: No module named 'tools.setup_helper'
      [end of output]
```
