### 2024-02-29 ISISNeutronMuon/quickBayes#89 @AnthonyLim23
how strange. It worked for me when it didnt before. It looks like we have the same version. What operating system are you using (maybe that makes a difference)?
```
(quickBayes2-dev) C:\Users\BTR75544\work\tmp>python -m pip install quickBayes
Collecting quickBayes
  Downloading quickBayes-1.0.0b21.tar.gz (2.2 MB)
     ---------------------------------------- 2.2/2.2 MB 6.3 MB/s eta 0:00:00
  Preparing metadata (setup.py) ... done
Requirement already satisfied: numpy>=1.12 in c:\users\btr75544\.conda\envs\quickbayes2-dev\lib\site-packages (from quickBayes) (1.26.4)
Requirement already satisfied: scipy in c:\users\btr75544\.conda\envs\quickbayes2-dev\lib\site-packages (from quickBayes) (1.12.0)
Requirement already satisfied: gofit in c:\users\btr75544\.conda\envs\quickbayes2-dev\lib\site-packages (from quickBayes) (0.4)
Building wheels for collected packages: quickBayes
  Building wheel for quickBayes (setup.py) ... done
  Created wheel for quickBayes: filename=quickBayes-1.0.0b21-py3-none-any.whl size=46202 sha256=5229cb2ef827426a61259cff6eb52aa60f57528450e7dc6d780edc12b8e9b621
  Stored in directory: c:\users\btr75544\appdata\local\pip\cache\wheels\cf\9b\ca\98ce560e5b6417de9cc9e12b0bc356af57f5b58f843b4153f7
Successfully built quickBayes
Installing collected packages: quickBayes
Successfully installed quickBayes-1.0.0b21
```
