### 2024-02-29 ISISNeutronMuon/quickBayes#89 @AnthonyLim23
I dont have access to a mac, but I will try in the next week on a Linux and see if that works. If not I will need to track down someone else to help me debug it. I am sorry this is taking awhile.
