+++ src/quickBayes/fit_engines/makefile.txt
@@ -2,3 +2,4 @@ fitting.fit_utils: join('fit_engines', 'fit_utils.py')
 fitting.fit_engine: join('fit_engines', 'fit_engine.py')
 fitting.scipy_engine: join('fit_engines', 'scipy_fit_engine.py')
 fitting.gofit_engine: join('fit_engines', 'gofit_engine.py')
+fitting.bumps_engine: join('fit_engines', 'bumps_fit_engine.py')
