### 2024-03-28 e7f3e35
Add conda build instructions to README
