### 2024-03-28 327015a
Expand README instructions for conda
