+++ .github/workflows/build_and_test.yml
@@ -22,19 +22,22 @@ jobs:
         uses: conda-incubator/setup-miniconda@v2.2.0
         with:
           miniforge-variant: Mambaforge
-          environment-file: qeb-dev-linux-py310.yml
-          activate-environment: qeb-dev-py310
+          environment-file: conda/recipe/qeb-ci-py310.yml
+          activate-environment: qeb-ci-py310
       - name: Setup Miniconda - Windows
         if: matrix.os == 'windows-latest'
         uses: conda-incubator/setup-miniconda@v2.2.0
         with:
           miniforge-variant: Mambaforge
-          environment-file: qeb-dev-win-py310.yml
-          activate-environment: qeb-dev-py310
+          environment-file: conda/recipe/qeb-ci-py310.yml
+          activate-environment: qeb-ci-py310
       - name: Build & Install
         run: |
-          python setup.py bdist_wheel
-          pip install --force-reinstall dist/quasielasticbayes-0.2.0-cp310-cp310-*.whl
+          cd conda/recipe
+          conda build .
+          echo ${PWD}
+          echo ${SRC_DIR}
+          conda install --use-local quasielasticbayes-0.2.0-*.tar.bz2
       - name: Test
         run: |
           pytest quasielasticbayes/test
