+++ conda/recipe/meta.yaml
@@ -6,7 +6,7 @@ package:
   version: {{ version }}
 
 source:
-  git_url: https://github.com/mantidproject/quasielasticbayes.git
+  git_url: git@github.com:mantidproject/quasielasticbayes.git
   git_rev: 621566baa1c229269a4679993a5b58ee2f2ce723
 
 build:
