+++ conda/recipe/conda_build_config.yaml
@@ -1,5 +1,7 @@
 compilers:
-  - m2w64-toolchain_win-64 # [win]
+  - m2w64-toolchain_win-64  # [win]
+  - gfortran                # [linux]
+  - libgfortran-ng          # [linux]
 
 python:
   - 3.10
