### 2024-03-28 dcb0306
Output built package to current working dir
