### 2024-03-28 79c7189
Add conda-verify to build and ci envs
