### 2024-03-28 574dcee
Update README to output package in cwd
