+++ conda/recipe/qeb-test-py310.yml
@@ -1,11 +1,9 @@
-name: qeb-ci-py310
+name: qeb-test-py310
 
 channels:
     - conda-forge
 
 dependencies:
-    - conda-build
-    - conda-verify
     - numpy
     - python=3.10.*
     - pytest
