+++ .github/workflows/build_and_test.yml
@@ -23,6 +23,7 @@ jobs:
           miniforge-variant: Mambaforge
           environment-file: conda/recipe/qeb-build-py310.yml
           activate-environment: qeb-build-py310
+          auto-activate-base: false
       - name: Build
         run: |
           cd conda/recipe
@@ -33,7 +34,8 @@ jobs:
           miniforge-variant: Mambaforge
           environment-file: conda/recipe/qeb-test-py310.yml
           activate-environment: qeb-test-py310
+          auto-activate-base: false
       - name: Install & Test
         run: |
-          conda install --use-local conda/recipe/*-64/quasielasticbayes-0.2.0-*.tar.bz2
+          conda install --force-reinstall --use-local conda/recipe/*-64/quasielasticbayes-0.2.0-*.tar.bz2
           pytest quasielasticbayes/test
