### 2024-03-28 bd64a09
Run tests in meta.yaml test stage
