+++ conda/recipe/qeb-ci-py310.yml
@@ -1,9 +1,11 @@
-name: qeb-test-py310
+name: qeb-ci-py310
 
 channels:
     - conda-forge
 
 dependencies:
+    - conda-build
+    - conda-verify
     - numpy
     - python=3.10.*
     - pytest
