### 2024-04-02 1d93408
Add Manifest file for generating source tarball
