+++ MANIFEST.in
@@ -1,0 +1,3 @@
+recursive-include . quasielasticbayes/*
+include *
+
