+++ MANIFEST.in
@@ -1,3 +1,2 @@
-recursive-include . quasielasticbayes/*
-include *
+recursive-include . *
 
