+++ .github/workflows/build_and_test.yml
@@ -16,7 +16,7 @@ jobs:
 
     steps:
       - name: Checkout Repository
-        uses: actions/checkout@v2
+        uses: actions/checkout@v4
       - name: Setup Build Env
         uses: conda-incubator/setup-miniconda@v3
         with:
