+++ conda/recipe/meta.yaml
@@ -6,8 +6,8 @@ package:
   version: {{ version }}
 
 source:
-  url: https://pypi.io/packages/source/{{ name[0] }}/{{ name }}/{{ name }}-{{ version }}.b4.tar.gz
-  sha256: 83a176633d6a4872ac843bc94b4ea42aae969cc6003e5d0b5e308604685fd98d
+  url: https://pypi.io/packages/source/{{ name[0] }}/{{ name }}/{{ name }}-{{ version }}.b5.tar.gz
+  sha256: 64eb64ea909eb40519519756f4763d07ab096454be6f7520f65756fbefe3841b
 
 build:
   script: {{ PYTHON }} setup.py install
