### 2024-04-03 0f918fb
Update MANIFEST.in to exclude build directory
