+++ MANIFEST.in
@@ -1,2 +1,3 @@
-recursive-include . *
+graft .
+prune build
 
