### 2024-04-03 78814b0
Add setuptools to meta yaml
