+++ conda/recipe/conda_build_config.yaml
@@ -1,6 +1,3 @@
 compilers:
   - m2w64-toolchain_win-64  # [win]
   - gfortran                # [linux]
-
-python:
-  - 3.10
