+++ conda/recipe/conda_build_config.yaml
@@ -1,3 +1,0 @@
-compilers:
-  - m2w64-toolchain_win-64  # [win]
-  - gfortran                # [linux]
