+++ conda/recipe/meta.yaml
@@ -16,7 +16,8 @@ build:
 
 requirements:
   build:
-    - {{ compilers }}
+    - m2w64-toolchain_win-64  # [win]
+    - gfortran                # [linux]
   host:
     - python
     - setuptools
