### 2024-04-03 e4d7e2e
Use pip install in build script
