### 2024-04-03 068190e
Add pre-commit to dev files
