+++ qeb-dev-win-py310.yml
@@ -8,4 +8,5 @@ dependencies:
     - m2w64-gcc-fortran=5.3.0
     - python=3.10.*
     - numpy=1.21.*
+    - pre-commit
     - pytest
