+++ qeb-dev-linux-py310.yml
@@ -8,4 +8,5 @@ dependencies:
     - libgfortran-ng=12.3.0
     - python=3.10.*
     - numpy=1.21.*
+    - pre-commit
     - pytest
