+++ .pre-commit-config.yaml
@@ -2,6 +2,8 @@ repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v3.4.0
     hooks:
+      - id: check-json
+        files: '\.(json)$'
       - id: pretty-format-json
         args:
           - "--autofix"
