### 2024-04-03 d69f006
Use relative tolerance for yprob check
