+++ quasielasticbayes/testing.py
@@ -47,13 +47,6 @@ def get_OS_precision():
         return 1
 
 
-def get_qlres_prob(ref):
-    if sys.platform == 'win32':
-        return ref
-    else:
-        return [-7.4106695e+04, -4.031369e+02, -3.96698e-1,  0.0]
-
-
 def get_qlse_prob(ref):
     if sys.platform == 'win32':
         return ref
