### 2024-04-04 12c12dd
Add OSX to conda build script
