+++ conda/recipe/meta.yaml
@@ -10,7 +10,7 @@ source:
   sha256: 64eb64ea909eb40519519756f4763d07ab096454be6f7520f65756fbefe3841b
 
 build:
-  skip: True  # [py<310 or py>310 or osx]
+  skip: True  # [py<310 or py>310]
   script: {{ PYTHON }} setup.py install
   number: 0
 
