### 2024-04-04 d1c6449
Reduce precision for MacOS tests
