+++ conda/recipe/meta.yaml
@@ -6,8 +6,9 @@ package:
   version: {{ version }}
 
 source:
-  url: https://pypi.io/packages/source/{{ name[0] }}/{{ name }}/{{ name }}-{{ version }}.b5.tar.gz
-  sha256: 64eb64ea909eb40519519756f4763d07ab096454be6f7520f65756fbefe3841b
+  path: ../../
+  #url: https://pypi.io/packages/source/{{ name[0] }}/{{ name }}/{{ name }}-{{ version }}.b5.tar.gz
+  #sha256: 64eb64ea909eb40519519756f4763d07ab096454be6f7520f65756fbefe3841b
 
 build:
   skip: True  # [py<310 or py>310]
