### 2024-04-05 33eb80a
Add manifest file for generating a source distribution
