+++ MANIFEST.in
@@ -1,0 +1,2 @@
+graft .
+prune build
