+++ README.md
@@ -1,8 +1,9 @@
 # quasielasticbayes
 
-This package provides a convenient set of Python wrappers
-for a set of routines used to perform Bayesian analysis
-on quasi-elastic neutron-scattering data.
+This package provides a convenient set of Python wrappers for a set of routines used to perform Bayesian analysis
+on quasi-elastic neutron-scattering data. The original Fortran code was written by Dr. Devinder Silvia, with 
+supervision from Dr. Spencer Howells, in the 1980's. The longevity of the package is owed to the efforts of
+Dr. Spencer Howells, and members of the Mantid Project team.
 
 ## Setup
 
