+++ README.md
@@ -7,92 +7,28 @@ Dr. Spencer Howells, and members of the Mantid Project team.
 
 ## Setup
 
-The simplest way to build the distribution is to use [conda](https://docs.conda.io/en/latest/miniconda.html) to create
-a separate environment to build the distribution.
-
-Create a minimal conda environment:
-
-### Windows - Python 3.10
+The simplest way to develop this package involves creating a [conda](https://docs.conda.io/en/latest/miniconda.html) 
+developer environment. To create a conda environment, run
 
 ```sh
-mamba env create -f conda/develop/qeb-win-py310.yml
+mamba env create -f conda/develop/qeb-<os>-py310.yml
+```
+
+To activate the environment, run
+
+```sh
 conda activate qeb-dev-py310
-```
-
-### Linux - Python 3.10
-
-```sh
-mamba env create -f conda/develop/qeb-linux-py310.yml
-conda activate qeb-dev-py310
-```
-
-### Windows - Python 3.8
-
-We currently rely on an external fortran compiler, `tdm64-gcc 4.6.1`, as the current code is sensitive
-to the compiler version. To install:
-
-- Download [tdm64-gcc-4.6.1.exe](https://sourceforge.net/projects/tdm-gcc/files/TDM-GCC%20Installer/Previous/1.1006.0/tdm64-gcc-4.6.1.exe/download) and install it to ``C:\MinGW64``
-- Download [gcc-4.6.1-tdm64-1-fortran.zip](https://sourceforge.net/projects/tdm-gcc/files/TDM-GCC%20Old%20Releases/TDM-GCC%204.6%20series/4.6.1-tdm64-1/gcc-4.6.1-tdm64-1-fortran.zip/download)
-- Right-click on ``gcc-4.6.1-tdm64-1-fortran.zip``, select "Extract All" and enter the location as ``C:\MinGW64``
-- Add ``C:\MinGW64\bin`` to your ``PATH`` environment variable ([instructions here](https://www.architectryan.com/2018/03/17/add-to-the-path-on-windows-10/))
-- Restart any terminal or powershell instances to capture the new environment variable settings
-
-Then create a developer conda environment:
-
-```sh
-mamba env create -f conda/develop/qeb-win-py38.yml
-conda activate qeb-dev-py38
-```
-
-### Linux - Python 3.8
-
-```sh
-mamba env create -f conda/develop/qeb-linux-py38.yml
-conda activate qeb-dev-py38
-```
-
-## Build and Test
-
-From the root of this repository:
-
-To build a wheel, run
-
-```sh
-python setup.py bdist_wheel
-```
-
-To install the wheel, run
-
-```sh
-pip install --force-reinstall dist/quasielasticbayes-0.2.0-cp310-cp310-*.whl
-```
-
-To run the tests
-
-```sh
-pytest quasielasticbayes/test
-```
-
-## Building for PyPi
-
-If this is your first time interacting with PyPi then please see [here](https://packaging.python.org/en/latest/tutorials/packaging-projects/#uploading-the-distribution-archives) for instructions of how to setup accounts and API tokens. 
-
-Once built the wheel can be uploaded using twine:
-
-```sh
-twine upload ./dist/name_of_wheel
 ```
 
 ## Building for conda-forge
 
-Enter the `conda/recipe` directory and activate the build environment
+To create and activate the build environment, run
 
 ```sh
-mamba env create -f conda/recipe/qeb-build-py310.yml
-conda activate qeb-build-py310
+mamba env create -f conda/recipe/qeb-build-py310.yml && conda activate qeb-build-py310
 ```
 
-To build the package, run
+Enter the `conda/recipe` directory. To build the package, run
 
 ```sh
 conda build . --output-folder .
[…]
