### 2024-04-05 4170a46
Simplify conda environments used for building the package
