+++ README.md
@@ -5,12 +5,28 @@ on quasi-elastic neutron-scattering data. The original Fortran code was written 
 supervision from Dr. Spencer Howells, in the 1980's. The longevity of the package is owed to the efforts of
 Dr. Spencer Howells, and members of the Mantid Project team.
 
-## Building a Conda package
+The package is available on [conda-forge](https://anaconda.org/conda-forge/quasielasticbayes) and [PyPi](https://pypi.org/project/quasielasticbayes).
+
+## Install the package
+
+To install this package from conda-forge, run
+
+```sh
+conda install -c conda-forge quasielasticbayes
+```
+
+To install this package from PyPi, run
+
+```sh
+pip install quasielasticbayes
+```
+
+## Develop: Building the Conda package
 
 To create and activate the build environment, run
 
 ```sh
-mamba env create -f conda/recipe/qeb-build-py310.yml && conda activate qeb-build-py310
+mamba env create -f conda/conda-build-py310.yml && conda activate qeb-conda-build-py310
 ```
 
 To build a Conda package, run
@@ -19,12 +35,14 @@ To build a Conda package, run
 conda build conda/recipe/ --output-folder .
 ```
 
-## Building a PyPi package
+At the end of the conda build step, the tests are run automatically to verify the package is working.
+
+## Develop: Building the PyPi package
 
 To create and activate the build environment, run
 
 ```sh
-mamba env create -f conda/develop/qeb-<os>-py310.yml && conda activate qeb-dev-py310
+mamba env create -f conda/pypi-build-py310.yml && conda activate qeb-pypi-build-py310
 ```
 
 To build a wheel and the source distribution tarball, run
@@ -33,21 +51,7 @@ To build a wheel and the source distribution tarball, run
 ./scripts/rebuild.sh
 ```
 
-## Run the Tests
-
-To create and activate the test environment, run
-
-```sh
-mamba env create -f conda/recipe/qeb-ci-py310.yml && conda activate qeb-ci-py310
-```
-
-To install a locally-built Conda package, run
-
-```sh
-conda install --use-local *-64/quasielasticbayes-*.tar.bz2
-```
-
-To install a locally-built PyPi wheel, run
+To install a locally-built PyPi wheel for testing, run
 
 ```sh
 pip install --force-reinstall dist/quasielasticbayes-*.whl
