+++ conda/develop/qeb-linux-py38.yml
@@ -1,12 +1,0 @@
-name: qeb-dev-py38
-
-channels:
-    - conda-forge
-
-dependencies:
-    - gfortran=12.3.0
-    - libgfortran-ng=12.3.0
-    - python=3.8.*
-    - numpy=1.16.*
-    - pre-commit
-    - pytest
