+++ conda/develop/qeb-win-py38.yml
@@ -1,11 +1,0 @@
-name: qeb-dev-py38
-
-channels:
-    - conda-forge
-
-dependencies:
-    # A fortran compiler must be provided separately as we require a very specific version of mingw
-    - python=3.8.*
-    - numpy=1.16.*
-    - pre-commit
-    - pytest
