### 2024-04-05 f0e66d3
Add pre-commit to conda build env for easier committing
