+++ conda/conda-build-py310.yml
@@ -8,3 +8,4 @@ dependencies:
     - conda-build
     - conda-verify
     - python=3.10.*
+    - pre-commit
