### 2024-04-07 97c2abc
Try read FFLAGS variable
