### 2024-04-07 baacbf4
Remove unwanted FFLAGS compiler options when using gfortran_osx-64 package
