### 2024-04-07 329d054
Remove flags from FFLAGS which cause numerical differences
