### 2024-04-07 4c1beda
Skip qlres test if OS is OSX
