### 2024-04-07 4bd9e23
Simplify compile flags
