### 2024-04-07 c77b6aa
Update test assertions to use assert_allclose
