+++ quasielasticbayes/test/qlres_test.py
@@ -4,8 +4,8 @@ import tempfile
 import unittest
 import numpy as np
 import sys
-from quasielasticbayes.testing import add_path, get_OS_precision, load_json
 
+from quasielasticbayes.testing import add_path, load_json
 from quasielasticbayes.QLres import qlres
 
 DATA_DIR = os.path.join(os.path.dirname(__file__), 'data')
@@ -37,12 +37,11 @@ class QLresTest(unittest.TestCase):
         with open(os.path.join(DATA_DIR, 'qlres', 'qlres-output-spec-0.json'), 'r') as fh:
             reference = load_json(fh)
         
-        dp = get_OS_precision()
         self.assertEqual(reference['nd'], nd)
-        np.testing.assert_almost_equal(reference['xout'], xout, decimal=dp)
-        np.testing.assert_almost_equal(reference['yout'], yout, decimal=dp)
-        np.testing.assert_almost_equal(reference['eout'], eout, decimal=dp)
-        np.testing.assert_almost_equal(reference['yfit'], yfit, decimal=dp)
+        np.testing.assert_allclose(reference['xout'], xout)
+        np.testing.assert_allclose(reference['yout'], yout)
+        np.testing.assert_allclose(reference['eout'], eout)
+        np.testing.assert_allclose(reference['yfit'], yfit, rtol=1e-3)
         np.testing.assert_allclose(reference['yprob'], yprob, rtol=1e-2)
 
 if __name__ == '__main__':
