### 2024-04-07 9506f73
Simplify unit tests to make it easier to read
