+++ quasielasticbayes/test/qlres_test.py
@@ -5,7 +5,7 @@ import unittest
 import numpy as np
 import sys
 
-from quasielasticbayes.testing import add_path, load_json
+from quasielasticbayes.testing import add_path, load_json, RELATIVE_TOLERANCE_FIT, RELATIVE_TOLERANCE_PROB
 from quasielasticbayes.QLres import qlres
 
 DATA_DIR = os.path.join(os.path.dirname(__file__), 'data')
@@ -20,7 +20,6 @@ class QLresTest(unittest.TestCase):
 
     @unittest.skipIf(sys.platform == "darwin", "Reading the json reference file causes an unexplained crash.")
     def test_qlres_minimal_input(self):
-        # reference inputs
         with open(os.path.join(DATA_DIR, 'qlres', 'qlres-input-spec-0.json'), 'r') as fh:
             inputs = load_json(fh)
 
@@ -41,8 +40,8 @@ class QLresTest(unittest.TestCase):
         np.testing.assert_allclose(reference['xout'], xout)
         np.testing.assert_allclose(reference['yout'], yout)
         np.testing.assert_allclose(reference['eout'], eout)
-        np.testing.assert_allclose(reference['yfit'], yfit, rtol=1e-3)
-        np.testing.assert_allclose(reference['yprob'], yprob, rtol=1e-2)
+        np.testing.assert_allclose(reference['yfit'], yfit, rtol=RELATIVE_TOLERANCE_FIT)
+        np.testing.assert_allclose(reference['yprob'], yprob, rtol=RELATIVE_TOLERANCE_PROB)
 
 if __name__ == '__main__':
     unittest.main()
