+++ conda/recipe/conda_build_config.yaml
@@ -1,3 +1,3 @@
 fortran_compiler:
-  - m2w64-toolchain_win-64 ==2.5.0     # [win]
-  - gfortran ==13.2.0                  # [linux or osx]
+  - m2w64-toolchain_win-64==2.5.0     # [win]
+  - gfortran==13.2.0                  # [linux or osx]
