### 2024-04-15 ISISNeutronMuon/quickBayes#89 @AnthonyLim23
Sorry it has taken so long for me to respond. I have asked around and I cannot find anyone who can reproduce your error. I just had a look at the tar file and it looks as expected (it includes the setup_helper.py)

```
CLRC+btr75544@NDLT1627 MINGW64 ~/Downloads/quickBayes-1.0.0b21/tools
$ ls
setup_helper.py

```

If you have any thoughts on what could be going wrong it would be greatly appreciated.
