### 2024-04-17 ISISNeutronMuon/quickBayes#89 @prashjet
> If you have any thoughts on what could be going wrong it would be greatly appreciated.

Unfortunately, I don't have any specific idea about what is going wrong, but I tried an install in a different conda environment (with Python 3.10.8) and got a different error message to before:

```
(quickBayes2) prash@air > python -m pip install quickBayes
Collecting quickBayes
  Using cached quickBayes-1.0.0b21.tar.gz (2.2 MB)
  Preparing metadata (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py egg_info did not run successfully.
  │ exit code: 1
  ╰─> [8 lines of output]
      Traceback (most recent call last):
        File "<string>", line 2, in <module>
        File "<pip-setuptools-caller>", line 34, in <module>
        File "/private/var/folders/n2/qrrgr98x5flg7lx88y_r9z0r0000gn/T/pip-install-4nccloy6/quickbayes_b7097a39198146069e10bd1d394708c0/setup.py", line 2, in <module>
          from tools.setup_helper import get_extensions
        File "/private/var/folders/n2/qrrgr98x5flg7lx88y_r9z0r0000gn/T/pip-install-4nccloy6/quickbayes_b7097a39198146069e10bd1d394708c0/tools/setup_helper.py", line 5, in <module>
          from Cython.Build import cythonize
      ModuleNotFoundError: No module named 'Cython'
      [end of output]

error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.
```

I then pip installed `cython` by hand, reattempted install of `quickBayes` then got the following error:

```
(quickBayes2) prash@air > python -m pip install quickBayes
Collecting quickBayes
  Using cached quickBayes-1.0.0b21.tar.gz (2.2 MB)
  Preparing metadata (setup.py) ... done
Requirement already satisfied: numpy>=1.12 in /opt/miniconda3/envs/quickBayes2/lib/python3.10/site-packages (from quickBayes) (1.23.4)
Requirement already satisfied: scipy in /opt/miniconda3/envs/quickBayes2/lib/python3.10/site-packages (from quickBayes) (1.9.3)
Collecting gofit (from quickBayes)
  Downloading gofit-0.4.tar.gz (74 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 74.7/74.7 kB 4.5 MB/s eta 0:00:00
  Preparing metadata (setup.py) ... done
Building wheels for collected packages: quickBayes, gofit
  Building wheel for quickBayes (setup.py) ... done
  Created wheel for quickBayes: filename=quickBayes-1.0.0b21-py3-none-any.whl size=46193 sha256=25f9daf3bc2d620d2c42206971bc397205c78ceaffcfe3eb0b939212d7dcd782
  Stored in directory: /Users/prash/Library/Caches/pip/wheels/f7/88/03/2ef2f8d25f22a6cf74615a23089706cc129634c2946e9d5b47
  Building wheel for gofit (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py bdist_wheel did not run successfully.
  │ exit code: 1
  ╰─> [4 lines of output]
      running bdist_wheel
      running build
      running build_ext
      error: [Errno 2] No such file or directory: 'cmake'
      [end of output]

  ERROR: Failed building wheel for gofit
  Running setup.py clean for gofit
Successfully built quickBayes
Failed to build gofit
ERROR: Could not build wheels for gofit, which is required to install pyproject.toml-based projects
```

I hope these are helpful!
