### 2024-04-17 ISISNeutronMuon/quickBayes#89 @AnthonyLim23
It will take me a little while, but I am going to do some tests on performance with its current form and without Cython (but using pyproject.toml). If it doesnt make much difference, then switching would be beneficial as it would simplify the github workflows and allow us to make a pure python wheel. This should fix the issues you have been having.
