### 2024-04-23 4278811
refs #89 move to pyproject.toml
