### 2024-04-23 c4abcc9
refs #89 move to pyproject.toml
