### 2024-04-23 f005f3c
refs #89 move to pyproject.toml
