### 2024-04-23 1f7b76c
refs #89 move to pyproject.toml
