+++ src/quickBayes/workflow/grid_search/grid_search_template.py
@@ -1,0 +1,264 @@
+from quickBayes.workflow.template import WorkflowTemplate
+from quickBayes.log_likelihood import loglikelihood
+from quickBayes.functions.base import BaseFitFunction
+
+from numpy import ndarray
+import numpy as np
+from abc import abstractmethod
+
+
+class Axis(object):
+    """
+    A simple object for holding the information
+    about an axis
+    """
+    def __init__(self, start, end, N, label='x'):
+        """
+        Create an axis
+        :param start: the first value for the axis
+        :param end: the last value for the axis
+        :param N: the number of values along the axis
+        :param label: the name of the axis
+        """
+        self._len = N
+        self._label = label
+        self._vals = np.linspace(start, end, N)
+
+    @property
+    def label(self) -> str:
+        """
+        Get the label for the axis
+        :return the label
+        """
+        return self._label
+
+    @property
+    def len(self):
+        """
+        Get the length of the axis
+        :return the length of the axis
+        """
+        return self._len
+
+    @property
+    def values(self):
+        """
+        Get the values used on the axis
+        :return the values from the axis
+        """
+        return self._vals
+
+
+class GridSearchTemplate(WorkflowTemplate):
+    """
+    A workflow for a grid search.
+    A grid search will do a series of fits for a range of
+    x and y values, that correspond to fixed parameters in
+    the fitting function.
+
+    The inherited class must include:
+    - _set_x_value
+    - _set_y_value
+    - N
+
+    The properties are:
+    - fit_engine
+    - get_grid
+    - get_x_axis
+    - get_y_axis
+
+    To add a fit engine:
+    - set_scipy_engine (scipy curve fit, recommended)
+    - set_gofit_engine (gofit)
+
+    Other methods:
+    - preprocess_data
+    - update_fit_engine
+    - update_function (call this one not the overwritten one)
+    - execute
+    - set_x_axis
+    - set_y_axis
+    - N
+    """
+    def __init__(self):
+        """
+        Set the results and error dicts for reporting
+        """
+        super().__init__()
+        self._x_axis = None
+        self._y_axis = None
+        self._grid = None
+
+    def set_x_axis(self, start: float, end: float,
+                   N: int, label: str) -> None:
+        """
+        Sets the x axis for the workflow
+        :param start: the first value on the x axis
+        :param end: the last value on the x axis
+        :param N: the number of values on the x axis
+        :param label: the x axis label
+        """
+        self._x_axis = Axis(start, end, N, label)
+
+    @property
+    def get_x_axis(self) -> Axis:
+        """
+        Get the x axis
+        :return the x axis object
+        """
+        return self._x_axis
+
+    def set_y_axis(self, start: float, end: float,
+                   N: int, label: str) -> None:
+        """
+        Sets the y axis for the workflow
+        :param start: the first value on the y axis
+        :param end: the last value on the y axis
+        :param N: the number of values on the y axis
+        :param label: the y axis label
+        """
+        self._y_axis = Axis(start, end, N, label)
+
+    @property
+    def get_y_axis(self) -> Axis:
+        """
+        Get the y axis
+        :return the y axis object
+        """
+        return self._y_axis
+
+    def _generate_grid(self) -> (ndarray, ndarray):
+        """
+        Creates a grid from the axis.
+        Need to have set the x and y axis first.
+        :return the X and Y values from np.meshgrid
+        """
+        if self._x_axis is None or self._y_axis is None:
+            raise ValueError("The x and/or y axis has"
+                             " not been set. Please use "
+                             "set_x_axis and set_y_axis.")
+
+        X, Y = np.meshgrid(self._x_axis.values,
+                           self._y_axis.values)
[…]
