### 2024-04-23 bf20f7c
refs #89 move to pyproject.toml
