### 2024-04-23 bb4a8f7
refs #89 move to pyproject.toml
