+++ docs/source/usage.rst
@@ -2,5 +2,5 @@ Test
 ====
 
 .. autofunction:: quickBayes.log_likelihood.loglikelihood
-.. autofunction:: quickBayes.fit_engines.scipy_fit_engine.ScipyFitEngine
+.. autofunction:: quickBayes.fit_engines.scipy_engine.ScipyFitEngine
 .. autofunction:: quickBayes.utils.parallel.parallel
