### 2024-04-23 6b47958
refs #89 move to pyproject.toml
