### 2024-04-23 0103169
refs #89 move to pyproject.toml
