+++ .github/workflows/release.yml
@@ -1,57 +1,37 @@
 name: Build and upload to PyPI
 
 on:
+  pull_request:
+    branches:
+      - 'main'
+ 
   release:
     types:
         - published
 
 jobs:
   build_wheels:
-    name: Build wheels on ${{ matrix.os }}
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-latest, windows-latest, macOS-latest]
+    name: Build wheels and src
+    runs-on: ubuntu-latest
     steps:
 
       - uses: actions/checkout@v3
         with:
           fetch-depth: "0"
 
+      - name: Install
+        run: python -m pip install build pytest --user
       - name: Build wheels
-        uses: pypa/cibuildwheel@v2.8.0
-        env:
-          # List of platforms to build on (incl. Python version)
-          CIBW_BUILD: cp3{8,9,10,11}-manylinux_x86_64 cp3{8,9,10,11}-win_amd64 cp3{8,9,10,11}-macosx_x86_64 cp3{8,9,10,11}-macosx_arm64
-
-          CIBW_BEFORE_BUILD: >
-             pip install numpy Cython scipy joblib
-
-          # Install test dependencies and run tests
-          CIBW_TEST_REQUIRES: pytest joblib
-          CIBW_TEST_COMMAND: >
-            pytest {project}/test
-      - uses: actions/upload-artifact@v3
+        run: python -m build --sdist --wheel
+      - name: Test wheel
+        run: python -m pytest
+      - name: upload wheel and src
+        uses: actions/upload-artifact@v3
         with:
-          path: ./wheelhouse/*.whl
-
-  build_sdist:
-    name: Build source distribution
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-
-      - name: Build sdist
-        run: |
-          pip install numpy Cython scipy joblib
-          python setup.py sdist
-
-      - uses: actions/upload-artifact@v3
-        with:
-          path: dist/*.tar.gz
+          path: dist/
 
   upload_pypi:
-    needs: [build_wheels, build_sdist]
+    needs: [build_wheels]
     runs-on: ubuntu-latest
     permissions:
       id-token: write
