### 2024-04-23 e7e809b
refs #89 move to pyproject.toml
