### 2024-04-23 a1d4cdf
refs #89 move to pyproject.toml
