### 2024-04-23 dc63eeb
refs #89 move to pyproject.toml
