+++ .github/workflows/release.yml
@@ -14,18 +14,17 @@ jobs:
     name: Build wheels and src
     runs-on: ubuntu-latest
     steps:
-
-      - uses: actions/checkout@v3
+      - name: checkout
+        uses: actions/checkout@v3
         with:
           fetch-depth: "0"
-
-      - name: Install
+      - name: install build
         run: python -m pip install build pytest --user
-      - name: Build wheels
+      - name: build wheel
         run: python -m build --sdist --wheel
-      - name: Test wheel
+      - name: test wheel
         run: python -m pytest
-      - name: upload wheel and src
+      - name: upload wheel
         uses: actions/upload-artifact@v3
         with:
           path: dist/
