### 2024-04-23 14e501e
refs #89 move to pyproject.toml
