+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = 'hatchling.build'
 [project]
 name = 'quickBayes'
 version = "1.0.0b22"
-requires-python = ">=3.8.1"
+requires-python = ">=3.7.1"
 dependencies = ['numpy', 'scipy', 'gofit']
 authors = [{name='Anthony Lim', email='anthony.lim@stfc.ac.uk'}]
 description = "A Bayesian fitting package used for model selection and grid searches of fits for neutron and muon data."
