### 2024-04-23 e0278c5
refs #89 move to pyproject.toml
