### 2024-04-23 a906416
refs #89 move to pyproject.toml
