+++ .github/workflows/release.yml
@@ -1,7 +1,6 @@
 name: Build and upload to PyPI
 
 on:
- on:
   push:
     branches:
     - 'main'
