### 2024-04-23 0ec7dca
refs #89 move to pyproject.toml
