+++ .github/workflows/merge.yml
@@ -1,0 +1,103 @@
+name: Build and upload to PyPI
+
+on:
+  push:
+    branches:
+    - 'main'
+  pull_request:
+    branches:
+      - 'main'
+  pull_request_review:
+    types: [submitted]
+permissions:
+  contents: write
+
+jobs:
+  update_version:
+    name: update version
+    if: github.event.review.state == 'approved'
+    runs-on: ubuntu-latest
+    steps:
+      - name: checkout
+        uses: actions/checkout@v4
+        with:
+          fetch-depth: "0"
+      - name: Update version on merge
+        run:  |
+            python tools/version.py "beta"
+      - name: Store changes
+        uses: actions/upload-artifact@v3
+        with:
+          name: pyproject
+          path: pyproject.toml
+
+  build:
+    name: Build wheels and src
+    runs-on: ubuntu-latest
+    needs: [update_version]
+    steps:
+      - name: checkout
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: "0"
+      - uses: actions/download-artifact@v3
+        with:
+          name: pyproject
+      - name: install build
+        run: python -m pip install build pytest --user
+      - name: build wheel
+        run: python -m build --sdist --wheel
+      - name: test wheel
+        run: python -m pytest
+      - name: upload wheel
+        uses: actions/upload-artifact@v3
+        with:
+          path: dist/
+
+  upload_pypi:
+    needs: [build]
+    runs-on: ubuntu-latest
+    permissions:
+      id-token: write
+    steps:
+      - name: mint API token
+        id: mint
+        uses: tschm/token-mint-action@v1.0.2
+
+      - uses: actions/download-artifact@v3
+        with:
+          name: artifact
+          path: dist
+
+      - name: Publish package distribution to PYPI
+        uses: pypa/gh-action-pypi-publish@release/v1
+        with:
+          password: ${{ steps.mint.outputs.api-token }}
+
+  update_setup:
+    needs: [upload_pypi]
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write
+    steps:
+      - name: checkout
+        uses: actions/checkout@v4
+        with:
+          fetch-depth: "0"
+          repository: ${{github.event.pull_request.head.repo.full_name }}
+          ref: ${{ github.event.pull_request.head.ref }}
+
+      - uses: actions/download-artifact@v3
+        with:
+          name: pyproject
+      - name: check
+        run: |
+          cat pyproject.toml
+          echo ${{github.event.pull_request.head.repo.full_name }}
+      - name: Commit on merge
+        run: |
+          git config user.name github-actions
+          git config user.email github-actions@github.com
+          git add pyproject.toml
+          git commit -m "update version number"
+          git push
