### 2024-04-23 7236df3
refs #89 move to pyproject.toml
