+++ .github/workflows/release.yml
@@ -1,13 +1,7 @@
 name: Build and upload to PyPI
 
 on:
-  push:
-    branches:
-    - 'main'
-  pull_request:
-    branches:
-      - 'main'
-  release:
+ release:
     types:
         - published
 
