### 2024-04-23 4f6976c
refs #89 move to pyproject.toml
