+++ .github/workflows/merge.yml
@@ -15,7 +15,6 @@ permissions:
 jobs:
   update_version:
     name: update version
-    if: github.event.review.state == 'approved'
     runs-on: ubuntu-latest
     steps:
       - name: checkout
