### 2024-04-23 05cdd3a
refs #89 move to pyproject.toml
