### 2024-04-23 5a01910
refs #89 move to pyproject.toml
