### 2024-04-23 3fa0bb1
refs #89 move to pyproject.toml
