### 2024-04-23 2fd35ea
refs #89 move to pyproject.toml
