+++ src/quickBayes/functions/makefile.txt
@@ -1,14 +1,0 @@
-functions.base: join('fit_functions', 'base.py')
-functions.BG: join('fit_functions', 'BG.py')
-functions.delta: join('fit_functions', 'delta_function.py')
-functions.lorentz: join('fit_functions', 'lorentz.py')
-functions.gaussian: join('fit_functions', 'gaussian.py')
-functions.composite: join('fit_functions', 'composite_fun.py')
-functions.convolution: join('fit_functions', 'conv_with_res.py')
-functions.qldata_function: join('fit_functions', 'qldata_function.py')
-functions.qe_function: join('fit_functions', 'quasielastic_function.py')
-functions.SE_fix: join('fit_functions', 'stretch_exp_fixed.py')
-functions.SE: join('fit_functions', 'stretch_exp.py')
-functions.qse_function: join('fit_functions', 'qse.py')
-functions.exp_decay: join('fit_functions', 'exp_decay.py')
-functions.qse_fixed: join('fit_functions', 'qse_fixed.py')
