### 2024-04-23 38c37bd
refs #89 move to pyproject.toml
