+++ .github/workflows/merge.yml
@@ -43,7 +43,7 @@ jobs:
         with:
           name: pyproject
       - name: install build
-        run: python -m pip install build pytest --user
+        run: python -m pip install build pytest
       - name: build wheel
         run: python -m build --sdist --wheel
       - name: test wheel
