### 2024-04-23 3fa300f
refs #89 move to pyproject.toml
