+++ .github/workflows/merge.yml
@@ -43,11 +43,9 @@ jobs:
         with:
           name: pyproject
       - name: install build
-        run: python -m pip install build pytest
+        run: python -m pip install build
       - name: build wheel
         run: python -m build --sdist --wheel
-      - name: test wheel
-        run: python -m pytest
       - name: upload wheel
         uses: actions/upload-artifact@v3
         with:
