+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = 'hatchling.build'
 
 [project]
 name = 'quickBayes'
-version = "0.2.0b0"
+version = "0.2.0b1"
 requires-python = ">=3.7.1"
 dependencies = ['numpy', 'scipy', 'gofit']
 authors = [{name='Anthony Lim', email='anthony.lim@stfc.ac.uk'}]
