### 2024-04-23 2bc94d5
refs #89 move to pyproject.toml
