### 2024-04-23 17b620a
refs #89 move to pyproject.toml
