+++ .github/workflows/release.yml
@@ -18,8 +18,6 @@ jobs:
         run: python -m pip install build pytest --user
       - name: build wheel
         run: python -m build --sdist --wheel
-      - name: test wheel
-        run: python -m pytest
       - name: upload wheel
         uses: actions/upload-artifact@v3
         with:
