### 2024-04-23 ISISNeutronMuon/quickBayes#89 @AnthonyLim23
@prashjet after some experimenting I have made the project a pure Python wheel, using pyproject..toml https://pypi.org/project/quickBayes/1.0.0b23/

Let me know if this solves your problem. If it does I will merge the changes into main.
