### 2024-04-25 ISISNeutronMuon/quickBayes#92 @AnthonyLim23
JOSS doc changes
As part of the JOSS review I will need to rewrite some of the docs. The plan is to have the following:

Introduction

- what is bayes? - priors, models, posteriors
- why use bayes? - to get the probability of a model and why this is important

Theory
- executive summary of the main equation and assumptions
- derive the main equation and highlight the assumptions
- Discussion on comparison to other methods ([AIC](https://en.wikipedia.org/wiki/Akaike_information_criterion), [BIC](https://en.wikipedia.org/wiki/Bayesian_information_criterion) and [nested sampling](https://en.wikipedia.org/wiki/Nested_sampling_algorithm) and MCMC)

Examples
- QENS data to identify the number of peaks present
- Identifying the number of decays in muon spectroscopy data
- Background identification (with creating a new workflow)
