### 2024-04-25 ISISNeutronMuon/quickBayes#93 @AnthonyLim23
API updates
The API docs need to be more detailed. At present they are not very informative.
