### 2024-04-30 #6230 @AnthonyLim23
I have outlined my plan for the new documentation https://github.com/ISISNeutronMuon/quickBayes/issues/92 I think this should cover all of the comments.

I also have an issue to do the API changes https://github.com/ISISNeutronMuon/quickBayes/issues/93
