### 2024-05-28 #6230 @AnthonyLim23
@osorensen I have written an introduction to Bayes and done a derivation of the main equation, using simplified notation (I have also extended it). I still need to clean it up a bit, I also need to write the comparison to other methods and the examples. Unfortunately, I have to do this around other things. I am sorry it is taking so long.

I think I have fixed the issue with not being able to install it https://github.com/ISISNeutronMuon/quickBayes/issues/89 and the fix  https://github.com/ISISNeutronMuon/quickBayes/pull/90
