+++ docs/source/theory.rst
@@ -1,0 +1,252 @@
+Theory
+======
+
+In this section the key equation for quickBayes is derived in detail.
+The first part is the origianl derivation as set out by Sivia.
+The second part is an extension of the derivation to include distinct lines.
+
+Original Derivation
+-------------------
+
+Define a model, :math:`M` as as sum of indistinguishable functions/lines
+
+.. math::
+
+    M = \sum_j A_j f(x, \theta)
+
+where, :math:`A_j` is amplitude, :math:`\theta` are parameters.
+
+The model posterior is
+
+.. math::
+
+   P(M|D) = P(D|M) \frac{(M)}{P(D)}
+
+where :math:`D` is the data and :math:`P` is the probability
+
+Assume a uniform prior :math:`P(M) = 1/N` for :math:`N` lines and :math:`P(D)` is the same for all models
+
+.. math::
+
+    P(M|D) \propto P(D|M)
+
+.. math::
+    :name: P(D|M)
+
+    P(D|M) \propto \int \mathrm{d}^N\theta P(D | \theta, M) P(\theta | M)
+
+Assume that we have a known space to investigate
+
+.. math::
+   :name: x
+
+   x_\mathrm{min} \le x \le x_\mathrm{max}
+
+
+.. math::
+   :name: A
+
+   0 \le A_j \le A_\mathrm{max} \label{A}
+
+So the normalization prior must be the volume of the hyper cube from :ref:`equation 2<x>` and :ref:`equation 3<A>` top hat function due to asssume flat prior -> volume
+
+.. math::
+   :name: P(theta|M)
+
+    P(\theta | M) = [(x_\mathrm{max} – x_\mathrm{min}) A_\mathrm{max}]^{-N}
+
+substituting :ref:`equation 4<P(theta|M)>` into :ref:`equation 1<P(D|M)>`
+
+.. math::
+
+   P(D|M) \propto [(x_\mathrm{max} – x_\mathrm{min}) A_\mathrm{max}]^{-N}\int \mathrm{d}^N\theta P(D|\theta, M)
+
+Assume that the data is subject to independent additive gaussian noise
+
+.. math::
+
+   P(D|\theta, M) \propto \exp\left(-\frac{\chi^2}{2}\right)
+
+where :math:`\chi^2` is the chi squared value and is a function of :math:`\theta`
+
+.. math::
+   :name: almost
+
+    P(D|M) \propto [(x_\mathrm{max} – x_\mathrm{min}) A_\mathrm{max}]^{-N}\int \mathrm{d}^N\theta \exp\left(-\frac{\chi^2}{2}\right)
+
+
+Assume a best fit exists with corresponding best fit parameters :math:`\theta_0` and a chi squared value of :math:`\chi_\mathrm{min}^2`
+Use a Taylor expansion
+
+.. math::
+
+   \chi^2 \approx \chi^2_\mathrm{min} + \frac{1}{2}[\underline{\theta} - \underline{\theta_0}]^\mathrm{T} \underline\nabla\ \underline\nabla \chi^2(\underline{\theta_0})[\underline{\theta} - \underline{\theta_0}]
+
+Hence
+
+.. math::
+   :name: Taylor
+
+   \int \mathrm{d}^N\theta \exp\left(-\frac{\chi^2}{2}\right) \approx \exp\left(-\frac{\chi^2_\mathrm{min}}{2}\right) \frac{(4\pi)^N}{\sqrt{(\mathrm{det}(\underline{\nabla} \ \underline{\nabla} \chi^2)) }}
+
[…]
