+++ docs/source/intro.rst
@@ -1,57 +1,24 @@
-Introduction
-============
+Welcome To quickBayes
+=====================
 
-The quickBayes package is an open source library for calculating Bayesian quantities in a short period of time, by making some assumptions. 
-The package is cross platform, supporting Windows, Mac OS and Linux. 
-This package has been developed by Anthony Lim from STFC’s ISIS Neutron and Muon facility. 
+The quickBayes package is an open source library for calculating Bayesian quantities in a short period of time, by making some assumptions.
+The package is cross platform, supporting Windows, Mac OS and Linux.
+This package has been developed by Anthony Lim from STFC’s ISIS Neutron and Muon facility.
 
-Background
-==========
+The quickBayes package was originally designed as a successor to the quasielasticbayes package.
+However, quickBayes has abstracted the key ideas to make a more generic package.
 
-A Bayesian analysis answers the question; what is the probability of my model given my data?
-The first part is easy to understand, if the model is a linear function and the data is a peak then the model is unlikely to be a good choice. 
-The second part is more subtle, the important part is; given my data.
-To help explain this we will consider two sets of data, they will have identical x and y values and the only difference will be the errors associated with the y values. 
-For this example we will use a single Gaussian with an amplitude of :math:`103`, mean of :math:`4.2` and sigma of :math:`0.9`.
-The following code was used to produce the data
+In science hypotheses are tested against data, to determine the underlying behavior of the system.
+These hypotheses can be in the form of a mathematical expression that originates from first principles (i.e. it has been derived) or is an approximation to other more complex mechanisms (e.g. semi-empirical methods).
+The quickBayes package is designed to test these hypotheses against the user’s data to identify the most likely.
+For example, in Quasi Elastic Neutron Scattering (QENS) the data can be represented by a summation of Lorentzian peaks (convoluted with a resolution function, see  \textbf{link to doc} for more detail), but the number of peaks is not obvious as shown by \textbf{insert figure}.
+The quickBayes packages is designed to make these decisions easier by calculating the probability of one, two or three peaks given the data.
 
-.. code-block:: python
-
-  from quickBayes.fit_functions.gaussian import Gaussin
-  import numpy as np
-
-  x = np.linspace(0, 10, 100)
-  noise = 1 + 0.1*(np.random,normal(0, 2, len(x)))
-  function = Gaussian()
-  ground_truth = function(x, 103, 4.2, .9)
-  y = ground_truth * noise
-  e1 = np.sqrt(y)
-  e2 = np.power(y, 0.1)
-     
-The two data sets are shown below. 
-
-.. figure:: /images/bayes_example_data.png
-   :alt: bayes_example_data.png
-   :width: 400px
-   :align: center
-
-Both data sets return the exact same fitting parameters with values :math:`102. \pm 3.` for the amplitude, :math:`4.20 \pm 0.03` for the mean and :math:`0.90 \pm 0.02` for sigma.
-The chi squared values are 0.003 for the first data set (:math:`\sqrt(y)`) and 0.04 for the second data set (with the smaller error bars).
-This is expected, but both results suggest a good fit. 
-
-The loglikelihood
-=================
-
-Alternatively, we can use the loglikelihood to calculate the log of the probability that the fit describes the data.
-Since it is a log of a probability the values will always be negative, with the most likely outcomes being closest to zero. 
-For the first data set the loglikelihood is :math:`-275.8` and for the second data set it is :math:`-7.0`.
-This means that the second data set is more likely to be represented by the model (fit), which is a consequence of the smaller error bars. 
-Furthermore, if we then consider the possibility of two gaussians the loglikelihoods are :math:`-58.2` for the first data set and :math:`-59.4` for the second data set. 
[…]

[…]
