+++ docs/source/install.rst
@@ -7,7 +7,7 @@ Installation
 ------------
 
 The quickBayes package is easy to install using code:`pip`.
-The library is available on `PyPi<https://pypi.org/project/quickBayes/#description>`_ and can be installed with pip:
+The library is available on `PyPi <https://pypi.org/project/quickBayes/#description>` and can be installed with pip:
 
 .. code-block:: python
 
@@ -16,10 +16,10 @@ The library is available on `PyPi<https://pypi.org/project/quickBayes/#descripti
 Reporting Issues
 ----------------
 
-To report an issue please create a `github issue<https://github.com/ISISNeutronMuon/quickBayes/issues>`_ with the details of the bug.
+To report an issue please create a `github issue <https://github.com/ISISNeutronMuon/quickBayes/issues>` with the details of the bug.
 
 
 Support
 -------
 
-Support will be provided via `github issues<https://github.com/ISISNeutronMuon/quickBayes/issues>`_, please outline the problem in a new issue.
+Support will be provided via `github issues <https://github.com/ISISNeutronMuon/quickBayes/issues>`, please outline the problem in a new issue.
