+++ docs/source/cf_methods.rst
@@ -64,6 +64,7 @@ Nested sampling
 AIC and BIC
 -----------
 
+`Mantid <https://mantidproject.org/>`_
 
 `AIC <https://en.wikipedia.org/wiki/Akaike_information_criterion>`
 
