+++ docs/source/dev/setup.rst
@@ -1,3 +1,5 @@
+.. _dev:
+
 Getting Started
 ===============
 
