+++ docs/source/workflows/index.rst
@@ -1,23 +1,0 @@
-.. quickBayes documentation master file, created by
-   sphinx-quickstart on Mon Apr  3 15:37:35 2023.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Workflows
-=========
-
-Contents
---------
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-   Workflows <workflows>
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
