+++ docs/source/install.rst
@@ -1,3 +1,5 @@
+.. _install:
+
 Getting Started
 ===============
 
