+++ docs/source/introduction.rst
@@ -1,3 +1,4 @@
+.. _stats:
 
 Introduction
 ============

[…]
