### 2024-06-11 #6230 @AnthonyLim23
The docs are working in progress but what I have so far are here: https://quickbayes--94.org.readthedocs.build/en/94/
The PR is https://github.com/ISISNeutronMuon/quickBayes/pull/94

I still need to do:
- [ ] add details on other Bayesian methods
- [ ] add examples (QENS, muon and background selection)
- [ ]  a general clean up of the wording (particularly in the theory section)

However, it should give a good indication of the new documentation and if it is suitable.
