### 2024-06-11 f10999d
refs #92 testing pynb inclusion
