+++ docs/source/conf.py
@@ -21,6 +21,7 @@ release = '1.0'
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
 extensions = ['sphinx.ext.autodoc',
+              'nbsphinx',
               'sphinx.ext.autosummary']
 
 templates_path = ['_templates']
