### 2024-06-11 b0b1dc7
refs #92 testing pynb inclusion
