+++ .readthedocs.yaml
@@ -19,6 +19,7 @@ python:
 
 
 sphinx:
+  configuration: docs/conf.py
   fail_on_warning: true
 
 formats:
