### 2024-06-11 9476ee7
refs #92 testing pynb inclusion
