+++ .readthedocs.yaml
@@ -19,7 +19,7 @@ python:
 
 
 sphinx:
-  configuration: docs/conf.py
+  configuration: docs/source/conf.py
   fail_on_warning: true
 
 formats:
