+++ .readthedocs.yaml
@@ -7,9 +7,6 @@ build:
   os: 'ubuntu-20.04'
   tools:
     python: 'mambaforge-22.9'
-  jobs:
-    pre_build:
-      - "jupyter-book config sphinx docs/"
 
 conda:
   environment: quickBayes-dev.yml
