+++ pyproject.toml
@@ -6,7 +6,7 @@ build-backend = 'hatchling.build'
 name = 'quickBayes'
 version = "1.0.0b23"
 requires-python = ">=3.7.1"
-dependencies = ['numpy', 'scipy', 'gofit']
+dependencies = ['numpy<2.0.0', 'scipy', 'gofit']
 authors = [{name='Anthony Lim', email='anthony.lim@stfc.ac.uk'}]
 description = "A Bayesian fitting package used for model selection and grid searches of fits for neutron and muon data."
 keywords=['bayesian', 'fitting', 'QENS', 'muons']
