### 2024-06-27 be48ef5
Fixing pip install (#90)
* #refs 89 first go at fixing pip install

* #refs 89 first go at fixing pip install

* #refs 89 first go at fixing pip install

* #refs 89 first go at fixing pip install

* #refs 89 first go at fixing pip install

* #refs 89 first go at fixing pip install

* #refs 89 first go at fixing pip install

* #refs 89 first go at fixing pip install

* #refs 89 first go at fixing pip install

* #refs 89 first go at fixing pip install

* refs #89 move to pyproject.toml

* refs #89 move to pyproject.toml

* refs #89 move to pyproject.toml

* refs #89 move to pyproject.toml

* ref
[…]
