+++ test/workflows/grid_search/quest_test.py
@@ -1,7 +1,7 @@
 import unittest
 from quickBayes.functions.qse_fixed import QSEFixFunction
 from quickBayes.functions.BG import LinearBG
-from quickBayes.workflow.qse_search import QSEGridSearch
+from quickBayes.workflow.grid_search.qse_grid_search import QSEGridSearch
 import numpy as np
 import os.path
 
