+++ src/quickBayes/workflow/model_selection/muon_decay.py
@@ -2,7 +2,7 @@ from quickBayes.functions.composite import CompositeFunction
 from quickBayes.functions.exp_decay import ExpDecay
 from quickBayes.utils.general import get_background_function
 from quickBayes.utils.crop_data import crop
-from quickBayes.workflow.model_template import ModelSelectionWorkflow
+from quickBayes.workflow.model_selection.template import ModelSelectionWorkflow
 from quickBayes.functions.base import BaseFitFunction
 from numpy import ndarray
 from typing import Dict, List
