+++ test/workflows/grid_search/gridSearchTemplate_test.py
@@ -1,10 +1,10 @@
 import unittest
-from quickBayes.workflow.grid_template import GridSearchTemplate
+from quickBayes.workflow.grid_search.template import GridSearchTemplate
 from quickBayes.functions.BG import FlatBG
 from quickBayes.functions.exp_decay import ExpDecay
-from quickBayes.test_helpers.workflows import (gen_grid_search_data,
-                                               FixedBG,
-                                               FixedComposite)
+from quickBayes.test_helpers.workflow_helper import (gen_grid_search_data,
+                                                     FixedBG,
+                                                     FixedComposite)
 
 
 class SimpleWorkflow(GridSearchTemplate):
