+++ src/quickBayes/workflow/model_selection/QSE.py
@@ -1,7 +1,7 @@
 from quickBayes.functions.qse_function import QSEFunction
 from quickBayes.utils.spline import spline
 from quickBayes.utils.general import get_background_function
-from quickBayes.workflow.model_template import ModelSelectionWorkflow
+from quickBayes.workflow.model_selection.template import ModelSelectionWorkflow
 from quickBayes.functions.base import BaseFitFunction
 from quickBayes.utils.crop_data import crop
 
