+++ src/quickBayes/makefile.txt
@@ -1,1 +1,0 @@
-log_likelihood: 'log_likelihood.py'
