+++ .readthedocs.yaml
@@ -7,6 +7,11 @@ build:
   os: 'ubuntu-20.04'
   tools:
     python: 'mambaforge-22.9'
+  jobs:
+    pre_build:
+      # Generate on-the-fly Sphinx configuration from Jupyter Book's _config.yml
+      - "jupyter-book config sphinx docs/"
+
 
 conda:
   environment: quickBayes-dev.yml
@@ -24,3 +29,4 @@ sphinx:
 
 formats:
   - pdf
+  - epub
