+++ .readthedocs.yaml
@@ -10,7 +10,7 @@ build:
   jobs:
     pre_build:
       # Generate on-the-fly Sphinx configuration from Jupyter Book's _config.yml
-      - "jupyter-book config sphinx docs/"
+      - "jupyter-book config sphinx docs/source/examples"
 
 
 conda:
