### 2024-06-27 ISISNeutronMuon/quickBayes#89 @prashjet
Hi @AnthonyLim23 - it didn't work for me I'm afraid!

```
(quickBayes2) prash@air > python -m pip install quickBayes
Collecting quickBayes
  Using cached quickbayes-1.0.0b23-py3-none-any.whl.metadata (1.1 kB)
Collecting gofit (from quickBayes)
  Using cached gofit-0.4.tar.gz (74 kB)
  Preparing metadata (setup.py) ... done
Requirement already satisfied: numpy in /opt/miniconda3/envs/quickBayes2/lib/python3.10/site-packages (from quickBayes) (1.23.4)
Requirement already satisfied: scipy in /opt/miniconda3/envs/quickBayes2/lib/python3.10/site-packages (from quickBayes) (1.9.3)
Using cached quickbayes-1.0.0b23-py3-none-any.whl (49 kB)
Building wheels for collected packages: gofit
  Building wheel for gofit (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py bdist_wheel did not run successfully.
  │ exit code: 1
  ╰─> [4 lines of output]
      running bdist_wheel
      running build
      running build_ext
      error: [Errno 2] No such file or directory: 'cmake'
      [end of output]

  ERROR: Failed building wheel for gofit
  Running setup.py clean for gofit
Failed to build gofit
ERROR: ERROR: Failed to build installable wheels for some pyproject.toml based projects (gofit)
```
