### 2024-06-27 ISISNeutronMuon/quickBayes#89 @AnthonyLim23
Ah ok, the output is much more helpful this time. The problem seems to be gofit: https://github.com/ralna/GOFit as it is a compiled code with a python interface. So I think the problem is just that it is not compatible with your system.

@jfowkes what do you think?
