+++ pyproject.toml
@@ -15,7 +15,7 @@ license = {text = 'BSD'}
 
 [project.optional-dependencies]
 docs = ["sphinx==7.3.7",
-	'jupyter',
+	'jupyter-book',
         "nbsphinx==0.9.4"]
 
 [project.urls]
