### 2024-06-28 ISISNeutronMuon/quickBayes#89 @jfowkes
@AnthonyLim23 not necessarily, GOFit is only precompiled for some systems, for others it tries to build GOFit from source (as is the case here) and that requires cmake amongst other dependencies, see the GOFit README here:
https://github.com/ralna/GOFit
