### 2024-06-28 ISISNeutronMuon/quickBayes#89 @jfowkes
@AnthonyLim23 probably best to make GoFit an optional install dependency?
