### 2024-06-28 ISISNeutronMuon/quickBayes#89 @AnthonyLim23
I was thinking that too. Would need to read up on how to do that and how to make the code cope with the possibility of it being missing.
