### 2024-06-28 ISISNeutronMuon/quickBayes#89 @AnthonyLim23
https://stackoverflow.com/questions/75490231/pip-pyproject-toml-can-optional-dependency-groups-require-other-optional-depend

and then I can use a base class for gofit that raises an error if it is not available, otherwise it will use the "implemented version".
