+++ docs/source/conf.py
@@ -8,9 +8,7 @@
 import pathlib
 import sys
 
-
 sys.path.insert(0, pathlib.Path(__file__).parents[2].resolve().as_posix())
-
 
 project = 'quickBayes'
 copyright = '2023, Anthony Lim'
