+++ .readthedocs.yaml
@@ -19,6 +19,7 @@ conda:
 
 python:
   install:
+    - requirements: docs/source/requirements.txt
     - method: pip
       path: .
 
