+++ quickBayes-dev.yml
@@ -13,6 +13,5 @@ dependencies:
   - Cython
   - gofit
   - sphinx
-  - jupyter-book
 - pip:
   - readthedocs-sphinx-ext
