+++ .readthedocs.yaml
@@ -19,7 +19,6 @@ conda:
 
 python:
   install:
-    - requirements: docs/source/requirements.txt
     - method: pip
       path: .
 
