### 2024-07-11 4bc7145
refs #92 some doc updates
