+++ docs/source/theory.rst
@@ -33,31 +33,31 @@ Assume a uniform prior :math:`P(M) = 1/N` for :math:`N` lines and :math:`P(D)` i
     P(M|D) \propto P(D|M)
 
 .. math::
-    :name: P(D|M)
+    :label: P(D|M)
 
     P(D|M) \propto \int \mathrm{d}^N\theta P(D | \theta, M) P(\theta | M)
 
 Assume that we have a known space to investigate
 
 .. math::
-   :name: x
+   :label: x
 
    x_\mathrm{min} \le x \le x_\mathrm{max}
 
 
 .. math::
-   :name: A
+   :label: A
 
    0 \le A_j \le A_\mathrm{max} \label{A}
 
-So the normalization prior must be the volume of the hyper cube from :ref:`equation 2<x>` and :ref:`equation 3<A>` top hat function due to asssume flat prior -> volume
+So the normalization prior must be the volume of the hyper cube from :math:numref:`x` and :math:numref:`A` top hat function due to asssume flat prior -> volume
 
 .. math::
-   :name: P(theta|M)
+   :label: P(theta|M)
 
     P(\theta | M) = [(x_\mathrm{max} – x_\mathrm{min}) A_\mathrm{max}]^{-N}
 
-substituting :ref:`equation 4<P(theta|M)>` into :ref:`equation 1<P(D|M)>`
+substituting :math:numref:`P(theta|M)` into :math:numref:`P(D|M)`
 
 .. math::
 
@@ -72,7 +72,7 @@ Assume that the data is subject to independent additive gaussian noise
 where :math:`\chi^2` is the chi squared value and is a function of :math:`\theta`
 
 .. math::
-   :name: almost
+   :label: almost
 
     P(D|M) \propto [(x_\mathrm{max} – x_\mathrm{min}) A_\mathrm{max}]^{-N}\int \mathrm{d}^N\theta \exp\left(-\frac{\chi^2}{2}\right)
 
@@ -87,22 +87,22 @@ Use a Taylor expansion
 Hence
 
 .. math::
-   :name: Taylor
+   :label: Taylor
 
    \int \mathrm{d}^N\theta \exp\left(-\frac{\chi^2}{2}\right) \approx \exp\left(-\frac{\chi^2_\mathrm{min}}{2}\right) \frac{(4\pi)^N}{\sqrt{(\mathrm{det}(\underline{\nabla} \ \underline{\nabla} \chi^2)) }}
 
 where :math:`\mathrm{det}(H) = \mathrm{det}(\underline{\nabla} \ \underline{\nabla} \chi^2))` is the determinant of the Hessian matrix :math:`H`.
-Substituting :ref:`equation 6<Taylor>` into :ref:`equation 5 <almost>` and for indistinguishable lines there are $N$ factorial possibilities
+Substituting :math:numref:`Taylor` into :math:numref:`almost` and for indistinguishable lines there are :math:`N` factorial possibilities
 
 .. math::
-   :name: sivia
+   :label: sivia
 
    P(D|M) \propto P(M|D) \propto \frac{N! (4\pi)^N }{[(x_\mathrm{max} - x_\mathrm{min})A_\mathrm{max}]^N \sqrt{H}} \exp\left(-\frac{\chi^2_0}{2}\right)
 
 Taking the logs and rearranging gives
 
 .. math::
-   :name: logs
+   :label: logs
 
    \log{[P(D|M)]} \propto \sum_{j=1}^{N}\log{(j)} +
    N\log{(4\pi)} - N\log{([x_\mathrm{max} - x_\mathrm{min}]A_\mathrm{max})} -
@@ -116,7 +116,7 @@ Including Unique Lines
[…]
