### 2024-08-05 #6230 @AnthonyLim23
@osorensen I'm sorry this is taking so long, I am having to do around other projects. I believe the only things left to do for the docs are:

- Tidy up the wording for the derivations
- The comparison to other techniques/models
- The examples (I do have the Python notebooks working)

I am hoping to get some time this week to do some work on it.
