### 2025-02-03 #6230 @AnthonyLim23
@osorensen, sorry I was off at the end of the year (paternity leave). The documentation changes are currently in review, I have chased the reviewer. However, I think it address all of the comments.

https://github.com/ISISNeutronMuon/quickBayes/pull/94
