### 2025-02-07 #6230 @osorensen
@JonasMoss, @mattpitkin, @prashjet, can you please take a look at the revised version of the JOSS submission and update your checklists? Let me know if you have any questions.
