### 2025-02-10 #6230 @AnthonyLim23
> Hi [@AnthonyLim23](https://github.com/AnthonyLim23), thanks for your work on the major updates to the documentation. I have just a few comments:
>
> * somewhere in the documentation, maybe in the ["Theory behind quickBayes"](https://quickbayes.readthedocs.io/en/latest/theory.html#theory), it should be noted that to find the best fit chi-squared value the code makes use of scipy or gofit, with references to both (I only see this mentioned in the muon example at the moment).
> * in the [developer docs section](https://quickbayes.readthedocs.io/en/latest/dev/setup.html#dev), the github clone URL is https://github.com/mantidproject/quickBayes.git rather than https://github.com/ISISNeutronMuon/quickBayes.git. Can this be fixed?
> * in the ["Theory behind quickBayes"](https://quickbayes.readthedocs.io/en/latest/theory.html) page is the a reference/references you could add to the original derivation of this (e.g., the [Sivia et al](https://www.sciencedirect.com/science/article/abs/pii/092145269290036R?via%3Dihub) paper if that is the correct one).
> * similarly, in the sections within the ["Bayesian Methods Comparison"](https://quickbayes.readthedocs.io/en/latest/cf/index.html) section, while you own descriptions are useful, it would be good to have some external links or references (even links to the wikipedia pages for MCMC, nested sampling, AIC and BIC would be useful).
> * in the AIC and BIC pages, there are a few cases where chi is not being render correctly - I expect there are just missing 's in the LaTeX.
> * in the equation defining the model [muon decay example](https://quickbayes.readthedocs.io/en/latest/examples/muon.html), I think the
>     N
>    should be above the summation symbol, i.e.,
>
>       ∑
>
>         j
>         =
>         1
>
>       N
>
>     A
>    rather than
>
>       ∑
>
>         j
>         =
>         1
>
>
>     N
>     A
> * in the muon and QENS examples, are there references that you can include to it's use?

I will make the changes this week.

The nested sampling does have a link to an article that gives a very good description, do I need to make it more obvious? I can add some links for MCMC, BIC and AIC.

I'm sorry, for the last point I don't understand what you mean by references to its use.
