### 2025-02-10 #6230 @mattpitkin
> I will make the changes this week.

Great, thanks.

> The nested sampling does have a link to an article that gives a very good description, do I need to make it more obvious? I can add some links for MCMC, BIC and AIC.

Sorry, yes, I missed the link there (even though its to a paper I'm an author on!)

> I'm sorry, for the last point I don't understand what you mean by references to its use.

I just mean, is there a paper in which the method is used for something like each example describes, so you can say, e.g.: "This example is a simplified version of the analysis used in Smith et al (2015) (with added DOI link!), which did X."
