### 2025-02-14 #6230 @AnthonyLim23
@mattpitkin I have a PR open with the changes https://github.com/ISISNeutronMuon/quickBayes/pull/97. This should get merged early next week, I will let you know when its done.

I have not added references for the muon or sin examples as they do not have any.
