### 2025-03-10 #6230 @osorensen
@mattpitkin, thanks a lot for your patience and for going through the changes made to this submission. Just a question: did you update [your checklist](https://github.com/openjournals/joss-reviews/issues/6230#issuecomment-1892775373) after the last round of revision?
