### 2025-03-10 #6230 @osorensen
@JonasMoss, @prashjet, can you please take a look at the revised version of the JOSS submission and update your checklists? Let me know if you have any questions.
