### 2025-03-10 #6230 @mattpitkin
@AnthonyLim23 I just have a few more comments about the paper before I sign-off the review. All suggestions are for the paragraph starting "Traditionally the calculation of the loglikelihood...". I would suggest:

* changing the use of log likelihood to "...marginal likelihood for a model, also known as the model evidence, ..." (the fact that you're calculating the log of it isn't really relevant to this part of the description).
* adding "... or Nested Sampling (Skilling, 2006)" to the end of the first sentence.
* adding a sentence following that along the lines of "Simpler and far less computationally demanding methods for model comparison exist, such as the _Bayesian Information Criterion_ or _Akaike Information Criterion_, which use the maximum likelihood and some basic method to prevent over-fitting by incorporating a penalty for more complex models."
* and then the subsequent sentence, it could be started with something like: "quickbayes attempts to provide a middle ground of a fast approximation whilst providing a more robust Bayesian approach to penalising complex models...".

Obviously, you should edit these statements for accuracy!

Nested sampling reference:

```bibtex
@article{10.1214/06-BA127,
author = {John Skilling},
title = {{Nested sampling for general Bayesian computation}},
volume = {1},
journal = {Bayesian Analysis},
number = {4},
publisher = {International Society for Bayesian Analysis},
pages = {833 -- 859},
keywords = {algorithm, annealing, Bayesian computation, evidence, marginal likelihood, Model selection, nest, phase change},
year = {2006},
doi = {10.1214/06-BA127},
URL = {https://doi.org/10.1214/06-BA127}
}
```
