### 2025-03-10 #6230 @mattpitkin
@AnthonyLim23 I think quastielasticbayes github reference will need to have some sort of year given. Seeing as you reference v0.2.0, and it looks like the v0.2.0 release was in this commit https://github.com/mantidproject/quasielasticbayes/commit/002136bca43bae906d79676233c141d112dfea4f in 2023, I suggest adding 2023 as the year.
