### 2025-03-10 #6230 @mattpitkin
I suppose I should extend this comment:

> changing the use of log likelihood to "...marginal likelihood for a model, also known as the model evidence, ..." (the fact that you're calculating the log of it isn't really relevant to this part of the description).

to the paper title, so that is instead: "quickBayes: An analytical approach to Bayesian marginal likelihoods". Would you be happy with this change? You might then need to add a sentence stating something like: "In previous iterations of this package, and in the prior quasielasticbayes package, the term loglikelihood was used as shorthand to describe the natural logarithm of the marginal likelihood or model evidence."
