### 2025-03-14 #6230 @JonasMoss
I think this looks fine. But I can't reproduce the first example section. I have Python 3.13.1,

>>> print(np.__version__)
1.26.4.

I'm running this file in a cloned package directory.

```{python}
from quickBayes.functions.composite import CompositeFunction
from quickBayes.utils.general import get_background_function
from quickBayes.workflow.model_selection.muon_decay import MuonExpDecay
import numpy as np
import matplotlib.pyplot as plt
import os

DATA_DIR = os.path.join("test", "shared", "data", "muon")

data_file = os.path.join(DATA_DIR, "muon_expdecay_2.npy")

sx, sy, se = np.loadtxt(data_file)

fig, ax = plt.subplots()
ax.errorbar(sx, sy, se, fmt="kx", label="data")
ax.set(xlabel="Time", ylabel="Asymmetry", title="Muon data")
plt.legend()

start_x = 0.16
end_x = 15.0
max_features = 4
results = {}
results_errors = {}

workflow = MuonExpDecay(results, results_errors)
workflow.preprocess_data(sx, sy, se, start_x, end_x)

BG = get_background_function("flat")
func = CompositeFunction()
func.add_function(BG)

lower, upper = func.get_bounds()
print("lower bounds", lower)
print("upper bounds", upper)

func.set_bounds([-0.1], [0.25])
print(func.get_bounds())

print(func.get_guess())

workflow.set_scipy_engine(func.get_guess(), *func.get_bounds())
func = workflow.execute(max_features, func, func.get_guess())

results, results_errors = workflow.get_parameters_and_errors
for key in results.keys():
    if "log" in key:
        print(key, results[key][0])

fit_engine = workflow.fit_engine
x1, y1, e1, _, _ = fit_engine.get_fit_values(0)

fit_engine = workflow.fit_engine
x2, y2, e2, _, _ = fit_engine.get_fit_values(1)

fig, ax = plt.subplots()
ax.errorbar(sx, sy, se, fmt="ok", label="data")
ax.set(xlabel="time (mu s)", ylabel="Asymmetry", title="input data")
ax.errorbar(x1, y1, e1, fmt="b--", label="1 exp")
ax.errorbar(x2, y2, e2, fmt="r--", label="2 exp")
ax.set(xlabel="Time (mu s)", ylabel="Asymmetry", title="Muon data")

plt.legend()

for key in results.keys():
    if "N2" in key and "log" not in key:
        print(key, results[key][0], results_errors[key][0])
```

(This is your example two with latex removed due to some bugs on *my side*. And modified file path)

* I get `raise LinAlgError("Singular matrix")`. Is this intended / ok? Maybe it should be caught and handled in some way then, or explained.
* The N4 loglikelihood doesnt appear in the results dictionary.

```{python}
>>> results.keys()
dict_keys(['N1:f1.BG constant', 'N1:f2.Amplitude', 'N1:f2.lambda', 'N1:loglikelihood', 'N2:f1.BG constant', 'N2:f2.Amplitude', 'N2:f2.lambda', 'N2:f3.Amplitude', 'N2:f3.lambda', 'N2:loglikelihood', 'N3:f1.BG constant', 'N3:f2.Amplitude', 'N3:f2.lambda', 'N3:f3.Amplitude', 'N3:f3.lambda', 'N3:f4.Amplitude', 'N3:f4.lambda', 'N3:loglikelihood', 'N4:f1.BG constant', 'N4:f2.Amplitude', 'N4:f2.lambda', 'N4:f3.Amplitude', 'N4:f3.lambda', 'N4:f4.Amplitude', 'N4:f4.lambda', 'N4:f5.Amplitude', 'N4:f5.lambda'])
```
