### 2025-03-14 #6230 @JonasMoss
I'd also add that your paper title is misleading; it would be better to give it a more specific name. Not all, or ever most, applications are not about muon scattering or physics, after all!
