### 2025-03-14 #6230 @osorensen
Thanks @JonasMoss!
@AnthonyLim23, can you please fix the [failing example](https://github.com/openjournals/joss-reviews/issues/6230#issuecomment-2724249223), and report back here when done? Also, based on [this](https://github.com/openjournals/joss-reviews/issues/6230#issuecomment-2724253863), perhaps edit the title to something more specifically related to the target applications of this submission?
