### 2025-03-14 #6230 @mattpitkin
I don't see a problem with the example using Python 3.10.16, NumPy 1.26.4 and SciPy 1.15.2. I'm guessing this is a Python 3.13 issue (in which case it might just be enough to say that the package doesn't support Python 3.13 yet).
