### 2025-03-14 #6230 @osorensen
> I don't see a problem with the example using Python 3.10.16, NumPy 1.26.4 and SciPy 1.15.2. I'm guessing this is a Python 3.13 issue (in which case it might just be enough to say that the package doesn't support Python 3.13 yet).

I agree.

@AnthonyLim23; I will read through the paper now and let you know if I have any further editorial comments.
