### 2025-03-14 #6230 @AnthonyLim23
@JonasMoss and @osorensen the package was originally developed for neutron and muon data, but there is nothing in the package that means the data has to be. The 3rd example is just some oscillating data (no application) https://quickbayes--107.org.readthedocs.build/en/107/examples/sin_wave.html . So I would argue that adding to the title that its just for neutron and muon data is misleading.
