### 2025-03-14 #6230 @AnthonyLim23
I have just merged in your change to the paper and defined MuSR as muon spectroscopy.
