### 2025-03-14 #6230 @osorensen
@AnthonyLim23, related to [this point](https://github.com/openjournals/joss-reviews/issues/6230#issuecomment-2724420555), could you consider updating [this line](https://github.com/ISISNeutronMuon/quickBayes/blob/53cb083040405e413be29026c1635595712f95f2/pyproject.toml#L8) so it contains an upper limit of the Python version? Alternatively, if the package is supposed to work for Python 3.13.1, could you please address [this point](https://github.com/openjournals/joss-reviews/issues/6230#issuecomment-2724249223)?
