### 2025-03-14 #6230 @AnthonyLim23
> [@AnthonyLim23](https://github.com/AnthonyLim23), related to [this point](https://github.com/openjournals/joss-reviews/issues/6230#issuecomment-2724420555), could you consider updating [this line](https://github.com/ISISNeutronMuon/quickBayes/blob/53cb083040405e413be29026c1635595712f95f2/pyproject.toml#L8) so it contains an upper limit of the Python version? Alternatively, if the package is supposed to work for Python 3.13.1, could you please address [this point](https://github.com/openjournals/joss-reviews/issues/6230#issuecomment-2724249223)?

The error message is from scipy (I have left the error messages to propagate from the individual fitting packages). That specific message is due to there being no solution for the fitting problem it has been presented with. I have noticed on other projects, changing anything in the setup of the machine (e.g. OS version, python version) can lead to the fitting package getting slightly different results. In this case the data really doesn't support 4 decays, so that slight change in values has lead to no solution being found. Does that make sense?

There are few things we could do (just one of these will fix it, but for the first 2 a bit of trial and error might be needed):

1. Change the fitting range in the example or locally
2. Change the start guess in the example or locally
3. Reduce the maximum number of decays to 3 in the example or locally

Which option would you prefer?
