### 2025-03-14 #6230 @AnthonyLim23
@osorensen the method is deterministic and the data is always the same (read from file).

I have tried on a few occasions but I have never been able to get to the bottom of why the changes in the computer setup lead to different results.

@JonasMoss could you please confirm that changing if the max_features to 3 works for you?
