### 2025-03-14 #6230 @mattpitkin
Just to add, I get the same results (running several times) with two different virtual environment setups: one had Python 3.10.16, NumPy 1.26.4 and SciPy 1.15.2 and the other had Python 3.10.13, NumPy 1.24.4 and SciPy 1.9.3 and the other had. These all gave:

```
N1:loglikelihood -161.625454386235
N2:loglikelihood -113.5483785235226
N3:loglikelihood -350.0282464689026
N4:loglikelihood -4960.03394765482
N2:f1.BG constant 0.025188493241681736 0.000597781052867103
N2:f2.Amplitude 0.1677717222740321 0.008152471917185926
N2:f2.lambda 3.8643082056440057 0.34286688648611474
N2:f3.Amplitude 0.09672961159122637 0.011012092317885828
N2:f3.lambda 1.0149729637410343 0.08170507902437864
```
