### 2025-03-14 #6230 @AnthonyLim23
I have just update the doc so that the max number of decays is 3. That should make it more reliable.
