### 2025-03-17 #6230 @AnthonyLim23
> [@AnthonyLim23](https://github.com/AnthonyLim23), I can confirm that the example posted by [@JonasMoss](https://github.com/JonasMoss) fails on my side with Python version 3.12.8 and the most recent numpy version, but that it succeeds when setting `max_features = 3`.
>
> I do however get different numerical results than [@mattpitkin](https://github.com/mattpitkin), as shown below:
>
> ```
> N1:loglikelihood -161.62545438577408
> N2:loglikelihood -113.5483784928295
> N3:loglikelihood -372.0374679183962
> N2:f1.BG constant 0.025188493403047678 0.0005977810349832843
> N2:f2.Amplitude 0.16777171979071276 0.008152472335655614
> N2:f2.lambda 3.864308359861787 0.342866903649225
> N2:f3.Amplitude 0.09672961659803334 0.01101209229000441
> N2:f3.lambda 1.0149729994034111 0.08170507664122857
> ```
>
> I will check with the other JOSS editors about whether this can be accepted for a deterministic algorithm, and get back to you.

Thank you.

It looks like the only real difference is when it uses 3 decays. Converting the log values to unnormalized probabilities (10^x), then for the 3 features it is zero. So I think its statistical noise and that is why the result is slightly different.

I am planning in the longer term to make this sort of thing more obvious by improving the calculation to give the Bayesian p value as described in https://jakevdp.github.io/blog/2015/08/07/frequentism-and-bayesianism-5-model-selection/
