### 2025-03-17 #6230 @osorensen
>It looks like the only real difference is when it uses 3 decays. Converting the log values to unnormalized probabilities (10^x), then for the 3 features it is zero. So I think its statistical noise and that is why the result is slightly different.

@AnthonyLim23, how can it be statistical noise if the algorithm is deterministic?
