### 2025-03-17 #6230 @AnthonyLim23
sorry, I meant numerical noise. Since it is an approximation for log_{10}(0)
