### 2025-03-17 #6230 @osorensen
> sorry, I meant numerical noise. Since it is an approximation for log_{10}(0)

I still don't follow. log_{10}(0) equal `-Inf`. Isn't it instead log_{10}(P), for some P>0?
