### 2025-03-17 #6230 @AnthonyLim23
The value that I have calculated is the `log_{10}` of the probability. Due to various approximations, which can be and will be improved in the future, and numerical noise `log_{10}(0) !=-Inf`, instead it is equal to a large negative number. In this case `-372`. I appreciate that about -300 does not seem large, but since it is an exponent a small decrease in the value results in a significant reduction in the probability.
