### 2025-03-17 #6230 @osorensen
But based on this excerpt from the documentation

```
The results and errors of the calculation can be obtained from the get_parameters_and_errors method. This returns two dictionaries, one for the values and one for the errors. The keys indicate the parameter, with the Nx showing how many features (decays) were used in the calculation. Within the results are the loglikelihoods (
), which are the logs of the unnormalized posterior probability. Hence, the most likely model is the one with the largest value. When the loglikelihood is calculated it does not take the background into account (other than being part of the
 value). Since the background is the same function in all of the models it just adds an offset to the loglikelihood, so it is neglected.

results, results_errors = workflow.get_parameters_and_errors
for key in results.keys():
    if 'log' in key:
        print(key, results[key][0])
N1:loglikelihood -161.625454386235
N2:loglikelihood -113.54837843974272
N3:loglikelihood -372.3725153229217
The results show that two decays are most likely, followed by one decay. The code includes a penalty for overfitting, this reduces the probability of a model that is overparameterized.
```

My reading is this:
- The log likelihood ratio of 2 compared to 1 decays is `-113.5 - (-161.6)`.
- The log likelihood ratio of 3 compared to 2 decays is `-372 - (-113.5)`.
- And so on.
Thus, the data seem to strongly suggest there are 2 decays.

Hence, there are two issues:
1. The probabilities inside the log are not normalized. Hence, exp(-372) is NOT the probability of 3 decays, and exp(-113) is NOT the probability of 2 decays. Instead, only the likelihood ratios are available. The probabilities underlying the likelihood computation would have to be normalized first, for this to work out.
2. It follows that the probability of 3 decays is not identically zero. It can in fact never be, unless your prior distribution puts zero probability on the event 3 decays. Hence, `-372` seems to be the logarithm of some extremely small, but nonzero number. The number may well be 0 to machine precision, but mathematically it is nonzero.

Agree?
