### 2025-03-17 #6230 @AnthonyLim23
Yes I do agree. For your 2 points:

1. It is why I want to improve the code in the future to provide the Bayesian ratios (`P_n/sum_j P_j`), which will provide a normalised probability of `n` features in . However, since the sum over all of the models being considered (`sum_j P_j`) is always the same then the most likely is the one with the largest `P_n` value (n is the number of features you are interested in). Which means when taking the `log_{10}` the most likely is the one with the largest value. I know this is all very hand wavy, but it is done this way to be consistent with the quasielasticbayes package.

2. I agree, but for the purposes of this discussion I would say that 3 decays is so unlikely and for all intents and purposes its zero (even though technically its non zero).
