### 2025-03-17 #6230 @osorensen
Thanks for the response @AnthonyLim23. Given that you have removed the fourth decay from the documentation, I understand that the issue pointed out by @JonasMoss should be fixed now.

@JonasMoss, if you think this is fine, could you please check off the last remaining box on your [review checklist](https://github.com/openjournals/joss-reviews/issues/6230#issuecomment-1893148188)?
