### 2025-03-18 #6230 @AnthonyLim23
> * ed release of your software, and list the version tag of the archived version here.
> * Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional

@osorensen I just wanted to double check, should I merge the PR for the paper before creating the archive?
