### 2025-03-18 #6230 @AnthonyLim23
The tag is v1.0.2 (I updated the version to prevent confusion with tags and release numbers) https://github.com/ISISNeutronMuon/quickBayes/releases/tag/v1.0.2

Zendo upload is https://zenodo.org/records/15045665

the DOI is 10.5281/zenodo.15045665
