### 2025-04-03 #6230 @AnthonyLim23
> [@AnthonyLim23](https://github.com/AnthonyLim23) – I'm helping [@crvernon](https://github.com/crvernon) out with this submission. Could you please merge this PR which has a couple of tiny fixes: [ISISNeutronMuon/quickBayes#112](https://github.com/ISISNeutronMuon/quickBayes/pull/112)

done
