### 2025-04-03 #6230 @danielskatz
👋 @arfon - the metadata seems to have encoding in it.  When this happens, I usually ask the authors to update the .md file to remove it, then reaccept the submission...
